Split out fake modules into extra files
- should be mostly backwards compatible
