+++ pyfakefs/fake_filesystem.py
@@ -14,24 +14,11 @@
 
 """A fake filesystem implementation for unit testing.
 
-:Includes:
-  * :py:class:`FakeFile`: Provides the appearance of a real file.
-  * :py:class:`FakeDirectory`: Provides the appearance of a real directory.
-  * :py:class:`FakeFilesystem`: Provides the appearance of a real directory
-    hierarchy.
-  * :py:class:`FakeOsModule`: Uses :py:class:`FakeFilesystem` to provide a
-    fake :py:mod:`os` module replacement.
-  * :py:class:`FakeIoModule`: Uses :py:class:`FakeFilesystem` to provide a
-    fake ``io`` module replacement.
-  * :py:class:`FakePathModule`:  Faked ``os.path`` module replacement.
-  * :py:class:`FakeFileOpen`:  Faked ``file()`` and ``open()`` function
-    replacements.
-
 :Usage:
 
->>> from pyfakefs import fake_filesystem
+>>> from pyfakefs import fake_filesystem, fake_os
 >>> filesystem = fake_filesystem.FakeFilesystem()
->>> os_module = fake_filesystem.FakeOsModule(filesystem)
+>>> os_module = fake_os.FakeOsModule(filesystem)
 >>> pathname = '/a/new/dir/new-file'
 
 Create a new file object, creating parent directory objects as needed:
@@ -94,18 +81,11 @@ True
 True
 """
 import errno
-import functools
 import heapq
-import inspect
-import io
-import locale
 import os
 import random
 import sys
-import traceback
-import uuid
 from collections import namedtuple, OrderedDict
-from contextlib import contextmanager
 from doctest import TestResults
 from enum import Enum
 from stat import (
@@ -116,9 +96,7 @@ from stat import (
     S_ISDIR,
     S_IFLNK,
     S_ISREG,
-    S_IFSOCK,
 )
-from types import ModuleType, TracebackType
 from typing import (
     List,
     Optional,
@@ -131,67 +109,18 @@ from typing import (
     AnyStr,
     overload,
     NoReturn,
-    ClassVar,
-    IO,
-    Iterator,
-    TextIO,
-    Type,
 )
-from pyfakefs.extra_packages import use_scandir
-from pyfakefs.fake_scandir import scandir, walk, ScanDirIter
+
+from pyfakefs import fake_file, fake_path, fake_io, fake_os, helpers, fake_open
+from pyfakefs.fake_file import AnyFileWrapper, AnyFile
 from pyfakefs.helpers import (
-    FakeStatResult,
-    BinaryBufferIO,
-    TextBufferIO,
     is_int_type,
-    is_byte_string,
-    is_unicode_string,
     make_string_path,
-    IS_PYPY,
     to_string,
     matching_string,
-    real_encoding,
-    now,
     AnyPath,
-    to_bytes,
+    AnyString,
 )
-from pyfakefs import __version__  # noqa: F401 for upwards compatibility
-
-PERM_READ = 0o400  # Read permission bit.
-PERM_WRITE = 0o200  # Write permission bit.
-PERM_EXE = 0o100  # Execute permission bit.
-PERM_DEF = 0o777  # Default permission bits.
-PERM_DEF_FILE = 0o666  # Default permission bits (regular file)
-PERM_ALL = 0o7777  # All permission bits.
-
-_OpenModes = namedtuple(
-    "_OpenModes",
-    "must_exist can_read can_write truncate append must_not_exist",
-)
-
-_OPEN_MODE_MAP = {
-    # mode name:(file must exist, can read, can write,
-    #            truncate, append, must not exist)
-    "r": (True, True, False, False, False, False),
-    "w": (False, False, True, True, False, False),
-    "a": (False, False, True, False, True, False),
-    "r+": (True, True, True, False, False, False),
-    "w+": (False, True, True, True, False, False),
-    "a+": (False, True, True, False, True, False),
-    "x": (False, False, True, False, False, True),
-    "x+": (False, True, True, False, False, True),
-}
-
-AnyFileWrapper = Union[
-    "FakeFileWrapper",
-    "FakeDirWrapper",
-    "StandardStreamWrapper",
-    "FakePipeWrapper",
-]
-
-AnyString = Union[str, bytes]
-
-AnyFile = Union["FakeFile", "FakeDirectory"]
 
 if sys.platform.startswith("linux"):
     # on newer Linux system, the default maximum recursion depth is 40
@@ -200,14 +129,6 @@ if sys.platform.startswith("linux"):
 else:
     # on MacOS and Windows, the maximum recursion depth is 32
     _MAX_LINK_DEPTH = 32
-
-NR_STD_STREAMS = 3
-if sys.platform == "win32":
-    USER_ID = 1
-    GROUP_ID = 1
-else:
-    USER_ID = os.getuid()
-    GROUP_ID = os.getgid()
 
 
 class OSType(Enum):
@@ -218,691 +139,29 @@ class OSType(Enum):
     WINDOWS = "windows"
 
 
-class PatchMode(Enum):
-    """Defines if patching shall be on, off, or in automatic mode.
-    Currently only used for `patch_open_code` option.
-    """
-
-    OFF = 1
-    AUTO = 2
-    ON = 3
-
-
-def set_uid(uid: int) -> None:
-    """Set the global user id. This is used as st_uid for new files
-    and to differentiate between a normal user and the root user (uid 0).
-    For the root user, some permission restrictions are ignored.
-
-    Args:
-        uid: (int) the user ID of the user calling the file system functions.
-    """
-    global USER_ID
-    USER_ID = uid
-
-
-def set_gid(gid: int) -> None:
-    """Set the global group id. This is only used to set st_gid for new files,
-    no permission checks are performed.
-
-    Args:
-        gid: (int) the group ID of the user calling the file system functions.
-    """
-    global GROUP_ID
-    GROUP_ID = gid
-
-
-def reset_ids() -> None:
-    """Set the global user ID and group ID back to default values."""
-    if sys.platform == "win32":
-        set_uid(1)
-        set_gid(1)
-    else:
-        set_uid(os.getuid())
-        set_gid(os.getgid())
-
-
-def is_root() -> bool:
-    """Return True if the current user is the root user."""
-    return USER_ID == 0
-
-
-class FakeLargeFileIoException(Exception):
-    """Exception thrown on unsupported operations for fake large files.
-    Fake large files have a size with no real content.
-    """
-
-    def __init__(self, file_path: str) -> None:
-        super(FakeLargeFileIoException, self).__init__(
-            "Read and write operations not supported for "
-            "fake large file: %s" % file_path
-        )
-
-
-def _copy_module(old: ModuleType) -> ModuleType:
-    """Recompiles and creates new module object."""
-    saved = sys.modules.pop(old.__name__, None)
-    new = __import__(old.__name__)
-    if saved is not None:
-        sys.modules[old.__name__] = saved
-    return new
-
-
-class FakeFile:
-    """Provides the appearance of a real file.
-
-    Attributes currently faked out:
-      * `st_mode`: user-specified, otherwise S_IFREG
-      * `st_ctime`: the time.time() timestamp of the file change time (updated
-        each time a file's attributes is modified).
-      * `st_atime`: the time.time() timestamp when the file was last accessed.
-      * `st_mtime`: the time.time() timestamp when the file was last modified.
-      * `st_size`: the size of the file
-      * `st_nlink`: the number of hard links to the file
-      * `st_ino`: the inode number - a unique number identifying the file
-      * `st_dev`: a unique number identifying the (fake) file system device
-        the file belongs to
-      * `st_uid`: always set to USER_ID, which can be changed globally using
-            `set_uid`
-      * `st_gid`: always set to GROUP_ID, which can be changed globally using
-            `set_gid`
-
-    .. note:: The resolution for `st_ctime`, `st_mtime` and `st_atime` in the
-        real file system depends on the used file system (for example it is
-        only 1s for HFS+ and older Linux file systems, but much higher for
-        ext4 and NTFS). This is currently ignored by pyfakefs, which uses
-        the resolution of `time.time()`.
-
-        Under Windows, `st_atime` is not updated for performance reasons by
-        default. pyfakefs never updates `st_atime` under Windows, assuming
-        the default setting.
-    """
-
-    stat_types = (
-        "st_mode",
-        "st_ino",
-        "st_dev",
-        "st_nlink",
-        "st_uid",
-        "st_gid",
-        "st_size",
-        "st_atime",
-        "st_mtime",
-        "st_ctime",
-        "st_atime_ns",
-        "st_mtime_ns",
-        "st_ctime_ns",
-    )
-
-    def __init__(
-        self,
-        name: AnyStr,
-        st_mode: int = S_IFREG | PERM_DEF_FILE,
-        contents: Optional[AnyStr] = None,
-        filesystem: Optional["FakeFilesystem"] = None,
-        encoding: Optional[str] = None,
-        errors: Optional[str] = None,
-        side_effect: Optional[Callable[["FakeFile"], None]] = None,
-    ):
-        """
-        Args:
-            name: Name of the file/directory, without parent path information
-            st_mode: The stat.S_IF* constant representing the file type (i.e.
-                stat.S_IFREG, stat.S_IFDIR), and the file permissions.
-                If no file type is set (e.g. permission flags only), a
-                regular file type is assumed.
-            contents: The contents of the filesystem object; should be a string
-                or byte object for regular files, and a dict of other
-                FakeFile or FakeDirectory objects wih the file names as
-                keys for FakeDirectory objects
-            filesystem: The fake filesystem where the file is created.
-            encoding: If contents is a unicode string, the encoding used
-                for serialization.
-            errors: The error mode used for encoding/decoding errors.
-            side_effect: function handle that is executed when file is written,
-                must accept the file object as an argument.
-        """
-        # to be backwards compatible regarding argument order, we raise on None
-        if filesystem is None:
-            raise ValueError("filesystem shall not be None")
-        self.filesystem: FakeFilesystem = filesystem
-        self._side_effect: Optional[Callable] = side_effect
-        self.name: AnyStr = name  # type: ignore[assignment]
-        self.stat_result = FakeStatResult(
-            filesystem.is_windows_fs, USER_ID, GROUP_ID, now()
-        )
-        if st_mode >> 12 == 0:
-            st_mode |= S_IFREG
-        self.stat_result.st_mode = st_mode
-        self.st_size: int = 0
-        self.encoding: Optional[str] = real_encoding(encoding)
-        self.errors: str = errors or "strict"
-        self._byte_contents: Optional[bytes] = self._encode_contents(contents)
-        self.stat_result.st_size = (
-            len(self._byte_contents) if self._byte_contents is not None else 0
-        )
-        self.epoch: int = 0
-        self.parent_dir: Optional[FakeDirectory] = None
-        # Linux specific: extended file system attributes
-        self.xattr: Dict = {}
-        self.opened_as: AnyString = ""
-
-    @property
-    def byte_contents(self) -> Optional[bytes]:
-        """Return the contents as raw byte array."""
-        return self._byte_contents
-
-    @property
-    def contents(self) -> Optional[str]:
-        """Return the contents as string with the original encoding."""
-        if isinstance(self.byte_contents, bytes):
-            return self.byte_contents.decode(
-                self.encoding or locale.getpreferredencoding(False),
-                errors=self.errors,
-            )
-        return None
-
-    @property
-    def st_ctime(self) -> float:
-        """Return the creation time of the fake file."""
-        return self.stat_result.st_ctime
-
-    @st_ctime.setter
-    def st_ctime(self, val: float) -> None:
-        """Set the creation time of the fake file."""
-        self.stat_result.st_ctime = val
-
-    @property
-    def st_atime(self) -> float:
-        """Return the access time of the fake file."""
-        return self.stat_result.st_atime
-
-    @st_atime.setter
-    def st_atime(self, val: float) -> None:
-        """Set the access time of the fake file."""
-        self.stat_result.st_atime = val
-
-    @property
-    def st_mtime(self) -> float:
-        """Return the modification time of the fake file."""
-        return self.stat_result.st_mtime
-
-    @st_mtime.setter
-    def st_mtime(self, val: float) -> None:
-        """Set the modification time of the fake file."""
-        self.stat_result.st_mtime = val
-
-    def set_large_file_size(self, st_size: int) -> None:
-        """Sets the self.st_size attribute and replaces self.content with None.
-
-        Provided specifically to simulate very large files without regards
-        to their content (which wouldn't fit in memory).
-        Note that read/write operations with such a file raise
-            :py:class:`FakeLargeFileIoException`.
-
-        Args:
-          st_size: (int) The desired file size
-
-        Raises:
-          OSError: if the st_size is not a non-negative integer,
-                   or if st_size exceeds the available file system space
-        """
-        self._check_positive_int(st_size)
-        if self.st_size:
-            self.size = 0
-        if self.filesystem:
-            self.filesystem.change_disk_usage(st_size, self.name, self.st_dev)
-        self.st_size = st_size
-        self._byte_contents = None
-
-    def _check_positive_int(self, size: int) -> None:
-        # the size should be an positive integer value
-        if not is_int_type(size) or size < 0:
-            self.filesystem.raise_os_error(errno.ENOSPC, self.name)
-
-    def is_large_file(self) -> bool:
-        """Return `True` if this file was initialized with size
-        but no contents.
-        """
-        return self._byte_contents is None
-
-    def _encode_contents(self, contents: Union[str, bytes, None]) -> Optional[bytes]:
-        if is_unicode_string(contents):
-            contents = bytes(
-                cast(str, contents),
-                self.encoding or locale.getpreferredencoding(False),
-                self.errors,
-            )
-        return cast(bytes, contents)
-
-    def set_initial_contents(self, contents: AnyStr) -> bool:
-        """Sets the file contents and size.
-           Called internally after initial file creation.
-
-        Args:
-            contents: string, new content of file.
-
-        Returns:
-            True if the contents have been changed.
-
-        Raises:
-              OSError: if the st_size is not a non-negative integer,
-                   or if st_size exceeds the available file system space
-        """
-        byte_contents = self._encode_contents(contents)
-        changed = self._byte_contents != byte_contents
-        st_size = len(byte_contents) if byte_contents else 0
-
-        current_size = self.st_size or 0
-        self.filesystem.change_disk_usage(
-            st_size - current_size, self.name, self.st_dev
-        )
-        self._byte_contents = byte_contents
-        self.st_size = st_size
-        self.epoch += 1
-        return changed
-
-    def set_contents(self, contents: AnyStr, encoding: Optional[str] = None) -> bool:
-        """Sets the file contents and size and increases the modification time.
-        Also executes the side_effects if available.
-
-        Args:
-          contents: (str, bytes) new content of file.
-          encoding: (str) the encoding to be used for writing the contents
-                    if they are a unicode string.
-                    If not given, the locale preferred encoding is used.
-
-        Returns:
-            True if the contents have been changed.
-
-        Raises:
-          OSError: if `st_size` is not a non-negative integer,
-                   or if it exceeds the available file system space.
-        """
-        self.encoding = real_encoding(encoding)
-        changed = self.set_initial_contents(contents)
-        if self._side_effect is not None:
-            self._side_effect(self)
-        return changed
-
-    @property
-    def size(self) -> int:
-        """Return the size in bytes of the file contents."""
-        return self.st_size
-
-    @size.setter
-    def size(self, st_size: int) -> None:
-        """Resizes file content, padding with nulls if new size exceeds the
-        old size.
-
-        Args:
-          st_size: The desired size for the file.
-
-        Raises:
-          OSError: if the st_size arg is not a non-negative integer
-                   or if st_size exceeds the available file system space
-        """
-
-        self._check_positive_int(st_size)
-        current_size = self.st_size or 0
-        self.filesystem.change_disk_usage(
-            st_size - current_size, self.name, self.st_dev
-        )
-        if self._byte_contents:
-            if st_size < current_size:
-                self._byte_contents = self._byte_contents[:st_size]
-            else:
-                self._byte_contents += b"\0" * (st_size - current_size)
-        self.st_size = st_size
-        self.epoch += 1
-
-    @property
-    def path(self) -> AnyStr:
-        """Return the full path of the current object."""
-        names: List[AnyStr] = []
-        obj: Optional[FakeFile] = self
-        while obj:
-            names.insert(0, matching_string(self.name, obj.name))  # type: ignore
-            obj = obj.parent_dir
-        sep = self.filesystem.get_path_separator(names[0])
-        if names[0] == sep:
-            names.pop(0)
-            dir_path = sep.join(names)
-            drive = self.filesystem.splitdrive(dir_path)[0]
-            # if a Windows path already starts with a drive or UNC path,
-            # no extra separator is needed
-            if not drive:
-                dir_path = sep + dir_path
-        else:
-            dir_path = sep.join(names)
-        return self.filesystem.absnormpath(dir_path)
-
-    if sys.version_info >= (3, 12):
-
-        @property
-        def is_junction(self) -> bool:
-            return self.filesystem.isjunction(self.path)
-
-    def __getattr__(self, item: str) -> Any:
-        """Forward some properties to stat_result."""
-        if item in self.stat_types:
-            return getattr(self.stat_result, item)
-        return super().__getattribute__(item)
-
-    def __setattr__(self, key: str, value: Any) -> None:
-        """Forward some properties to stat_result."""
-        if key in self.stat_types:
-            return setattr(self.stat_result, key, value)
-        return super().__setattr__(key, value)
-
-    def __str__(self) -> str:
-        return "%r(%o)" % (self.name, self.st_mode)
-
-
-class FakeNullFile(FakeFile):
-    def __init__(self, filesystem: "FakeFilesystem") -> None:
-        devnull = "nul" if filesystem.is_windows_fs else "/dev/null"
-        super(FakeNullFile, self).__init__(devnull, filesystem=filesystem, contents="")
-
-    @property
-    def byte_contents(self) -> bytes:
-        return b""
-
-    def set_initial_contents(self, contents: AnyStr) -> bool:
-        return False
-
-
-class FakeFileFromRealFile(FakeFile):
-    """Represents a fake file copied from the real file system.
-
-    The contents of the file are read on demand only.
-    """
-
-    def __init__(
-        self,
-        file_path: str,
-        filesystem: "FakeFilesystem",
-        side_effect: Optional[Callable] = None,
-    ) -> None:
-        """
-        Args:
-            file_path: Path to the existing file.
-            filesystem: The fake filesystem where the file is created.
-
-        Raises:
-            OSError: if the file does not exist in the real file system.
-            OSError: if the file already exists in the fake file system.
-        """
-        super().__init__(
-            name=os.path.basename(file_path),
-            filesystem=filesystem,
-            side_effect=side_effect,
-        )
-        self.contents_read = False
-
-    @property
-    def byte_contents(self) -> Optional[bytes]:
-        if not self.contents_read:
-            self.contents_read = True
-            with io.open(self.file_path, "rb") as f:
-                self._byte_contents = f.read()
-        # On MacOS and BSD, the above io.open() updates atime on the real file
-        self.st_atime = os.stat(self.file_path).st_atime
-        return self._byte_contents
-
-    def set_contents(self, contents, encoding=None):
-        self.contents_read = True
-        super(FakeFileFromRealFile, self).set_contents(contents, encoding)
-
-    def is_large_file(self):
-        """The contents are never faked."""
-        return False
-
-
-class FakeDirectory(FakeFile):
-    """Provides the appearance of a real directory."""
-
-    def __init__(
-        self,
-        name: str,
-        perm_bits: int = PERM_DEF,
-        filesystem: Optional["FakeFilesystem"] = None,
-    ):
-        """
-        Args:
-            name:  name of the file/directory, without parent path information
-            perm_bits: permission bits. defaults to 0o777.
-            filesystem: if set, the fake filesystem where the directory
-                is created
-        """
-        FakeFile.__init__(self, name, S_IFDIR | perm_bits, "", filesystem=filesystem)
-        # directories have the link count of contained entries,
-        # including '.' and '..'
-        self.st_nlink += 1
-        self._entries: Dict[str, AnyFile] = {}
-
-    def set_contents(self, contents: AnyStr, encoding: Optional[str] = None) -> bool:
-        raise self.filesystem.raise_os_error(errno.EISDIR, self.path)
-
-    @property
-    def entries(self) -> Dict[str, FakeFile]:
-        """Return the list of contained directory entries."""
-        return self._entries
-
-    @property
-    def ordered_dirs(self) -> List[str]:
-        """Return the list of contained directory entry names ordered by
-        creation order.
-        """
-        return [
-            item[0]
-            for item in sorted(self._entries.items(), key=lambda entry: entry[1].st_ino)
-        ]
-
-    def add_entry(self, path_object: FakeFile) -> None:
-        """Adds a child FakeFile to this directory.
-
-        Args:
-            path_object: FakeFile instance to add as a child of this directory.
-
-        Raises:
-            OSError: if the directory has no write permission (Posix only)
-            OSError: if the file or directory to be added already exists
-        """
-        if (
-            not is_root()
-            and not self.st_mode & PERM_WRITE
-            and not self.filesystem.is_windows_fs
-        ):
-            raise OSError(errno.EACCES, "Permission Denied", self.path)
-
-        path_object_name: str = to_string(path_object.name)
-        if path_object_name in self.entries:
-            self.filesystem.raise_os_error(errno.EEXIST, self.path)
-
-        self._entries[path_object_name] = path_object
-        path_object.parent_dir = self
-        if path_object.st_ino is None:
-            self.filesystem.last_ino += 1
-            path_object.st_ino = self.filesystem.last_ino
-        self.st_nlink += 1
-        path_object.st_nlink += 1
-        path_object.st_dev = self.st_dev
-        if path_object.st_nlink == 1:
-            self.filesystem.change_disk_usage(
-                path_object.size, path_object.name, self.st_dev
-            )
-
-    def get_entry(self, pathname_name: str) -> AnyFile:
-        """Retrieves the specified child file or directory entry.
-
-        Args:
-            pathname_name: The basename of the child object to retrieve.
-
-        Returns:
-            The fake file or directory object.
-
-        Raises:
-            KeyError: if no child exists by the specified name.
-        """
-        pathname_name = self._normalized_entryname(pathname_name)
-        return self.entries[to_string(pathname_name)]
-
-    def _normalized_entryname(self, pathname_name: str) -> str:
-        if not self.filesystem.is_case_sensitive:
-            matching_names = [
-                name for name in self.entries if name.lower() == pathname_name.lower()
-            ]
-            if matching_names:
-                pathname_name = matching_names[0]
-        return pathname_name
-
-    def remove_entry(self, pathname_name: str, recursive: bool = True) -> None:
-        """Removes the specified child file or directory.
-
-        Args:
-            pathname_name: Basename of the child object to remove.
-            recursive: If True (default), the entries in contained directories
-                are deleted first. Used to propagate removal errors
-                (e.g. permission problems) from contained entries.
-
-        Raises:
-            KeyError: if no child exists by the specified name.
-            OSError: if user lacks permission to delete the file,
-                or (Windows only) the file is open.
-        """
-        pathname_name = self._normalized_entryname(pathname_name)
-        entry = self.get_entry(pathname_name)
-        if self.filesystem.is_windows_fs:
-            if entry.st_mode & PERM_WRITE == 0:
-                self.filesystem.raise_os_error(errno.EACCES, pathname_name)
-            if self.filesystem.has_open_file(entry):
-                self.filesystem.raise_os_error(errno.EACCES, pathname_name)
-        else:
-            if not is_root() and (
-                self.st_mode & (PERM_WRITE | PERM_EXE) != PERM_WRITE | PERM_EXE
-            ):
-                self.filesystem.raise_os_error(errno.EACCES, pathname_name)
-
-        if recursive and isinstance(entry, FakeDirectory):
-            while entry.entries:
-                entry.remove_entry(list(entry.entries)[0])
-        elif entry.st_nlink == 1:
-            self.filesystem.change_disk_usage(-entry.size, pathname_name, entry.st_dev)
-
-        self.st_nlink -= 1
-        entry.st_nlink -= 1
-        assert entry.st_nlink >= 0
-
-        del self.entries[to_string(pathname_name)]
-
-    @property
-    def size(self) -> int:
-        """Return the total size of all files contained
-        in this directory tree.
-        """
-        return sum([item[1].size for item in self.entries.items()])
-
-    @size.setter
-    def size(self, st_size: int) -> None:
-        """Setting the size is an error for a directory."""
-        raise self.filesystem.raise_os_error(errno.EISDIR, self.path)
-
-    def has_parent_object(self, dir_object: "FakeDirectory") -> bool:
-        """Return `True` if dir_object is a direct or indirect parent
-        directory, or if both are the same object."""
-        obj: Optional[FakeDirectory] = self
-        while obj:
-            if obj == dir_object:
-                return True
-            obj = obj.parent_dir
-        return False
-
-    def __str__(self) -> str:
-        description = super(FakeDirectory, self).__str__() + ":\n"
-        for item in self.entries:
-            item_desc = self.entries[item].__str__()
-            for line in item_desc.split("\n"):
-                if line:
-                    description = description + "  " + line + "\n"
-        return description
-
-
-class FakeDirectoryFromRealDirectory(FakeDirectory):
-    """Represents a fake directory copied from the real file system.
-
-    The contents of the directory are read on demand only.
-    """
-
-    def __init__(
-        self,
-        source_path: AnyPath,
-        filesystem: "FakeFilesystem",
-        read_only: bool,
-        target_path: Optional[AnyPath] = None,
-    ):
-        """
-        Args:
-            source_path: Full directory path.
-            filesystem: The fake filesystem where the directory is created.
-            read_only: If set, all files under the directory are treated
-                as read-only, e.g. a write access raises an exception;
-                otherwise, writing to the files changes the fake files
-                only as usually.
-            target_path: If given, the target path of the directory,
-                otherwise the target is the same as `source_path`.
-
-        Raises:
-            OSError: if the directory does not exist in the real file system
-        """
-        target_path = target_path or source_path
-        real_stat = os.stat(source_path)
-        super(FakeDirectoryFromRealDirectory, self).__init__(
-            name=to_string(os.path.split(target_path)[1]),
-            perm_bits=real_stat.st_mode,
-            filesystem=filesystem,
-        )
-
-        self.st_ctime = real_stat.st_ctime
-        self.st_atime = real_stat.st_atime
-        self.st_mtime = real_stat.st_mtime
-        self.st_gid = real_stat.st_gid
-        self.st_uid = real_stat.st_uid
-        self.source_path = source_path  # type: ignore
-        self.read_only = read_only
-        self.contents_read = False
-
-    @property
-    def entries(self) -> Dict[str, FakeFile]:
-        """Return the list of contained directory entries, loading them
-        if not already loaded."""
-        if not self.contents_read:
-            self.contents_read = True
-            base = self.path
-            for entry in os.listdir(self.source_path):
-                source_path = os.path.join(self.source_path, entry)
-                target_path = os.path.join(base, entry)  # type: ignore
-                if os.path.islink(source_path):
-                    self.filesystem.add_real_symlink(source_path, target_path)
-                elif os.path.isdir(source_path):
-                    self.filesystem.add_real_directory(
-                        source_path, self.read_only, target_path=target_path
-                    )
-                else:
-                    self.filesystem.add_real_file(
-                        source_path, self.read_only, target_path=target_path
-                    )
-        return self._entries
-
-    @property
-    def size(self) -> int:
-        # we cannot get the size until the contents are loaded
-        if not self.contents_read:
-            return 0
-        return super(FakeDirectoryFromRealDirectory, self).size
-
-    @size.setter
-    def size(self, st_size: int) -> None:
-        raise self.filesystem.raise_os_error(errno.EISDIR, self.path)
+# definitions for backwards compatibility
+FakeFile = fake_file.FakeFile
+FakeNullFile = fake_file.FakeNullFile
+FakeFileFromRealFile = fake_file.FakeFileFromRealFile
+FakeDirectory = fake_file.FakeDirectory
+FakeDirectoryFromRealDirectory = fake_file.FakeDirectoryFromRealDirectory
+FakeFileWrapper = fake_file.FakeFileWrapper
+StandardStreamWrapper = fake_file.StandardStreamWrapper
+FakeDirWrapper = fake_file.FakeDirWrapper
+FakePipeWrapper = fake_file.FakePipeWrapper
+
+FakePathModule = fake_path.FakePathModule
+FakeOsModule = fake_os.FakeOsModule
+FakeFileOpen = fake_open.FakeFileOpen
+FakeIoModule = fake_io.FakeIoModule
+if sys.platform != "win32":
+    FakeFcntlModule = fake_io.FakeFcntlModule
+PatchMode = fake_io.PatchMode
+
+is_root = helpers.is_root
+set_uid = helpers.set_uid
+set_gid = helpers.set_gid
+reset_ids = helpers.reset_ids
 
 
 class FakeFilesystem:
@@ -964,7 +223,7 @@ class FakeFilesystem:
         self._cwd = ""
 
         # We can't query the current value without changing it:
-        self.umask = os.umask(0o22)
+        self.umask: int = os.umask(0o22)
         os.umask(self.umask)
 
         # A list of open file objects. Their position in the list is their
@@ -978,7 +237,7 @@ class FakeFilesystem:
         self.mount_points: Dict[AnyString, Dict] = OrderedDict()
         self._add_root_mount_point(total_size)
         self._add_standard_streams()
-        self.dev_null = FakeNullFile(self)
+        self.dev_null: Any = FakeNullFile(self)
         # set from outside if needed
         self.patch_open_code = PatchMode.OFF
         self.shuffle_listdir_results = False
@@ -1235,7 +494,7 @@ class FakeFilesystem:
                 current_dir = cast(FakeDirectory, directory)
 
         for new_dir in new_dirs:
-            new_dir.st_mode = S_IFDIR | PERM_DEF
+            new_dir.st_mode = S_IFDIR | helpers.PERM_DEF
 
         return current_dir
 
@@ -1459,13 +718,15 @@ class FakeFilesystem:
             path, follow_symlinks, allow_fd=True, check_owner=True
         )
         if self.is_windows_fs and not force_unix_mode:
-            if mode & PERM_WRITE:
+            if mode & helpers.PERM_WRITE:
                 file_object.st_mode = file_object.st_mode | 0o222
             else:
                 file_object.st_mode = file_object.st_mode & 0o777555
         else:
-            file_object.st_mode = (file_object.st_mode & ~PERM_ALL) | (mode & PERM_ALL)
-        file_object.st_ctime = now()
+            file_object.st_mode = (file_object.st_mode & ~helpers.PERM_ALL) | (
+                mode & helpers.PERM_ALL
+            )
+        file_object.st_ctime = helpers.now()
 
     def utime(
         self,
@@ -1512,7 +773,7 @@ class FakeFilesystem:
             file_object.st_atime_ns = ns[0]
             file_object.st_mtime_ns = ns[1]
         else:
-            current_time = now()
+            current_time = helpers.now()
             file_object.st_atime = current_time
             file_object.st_mtime = current_time
 
@@ -2302,10 +1563,10 @@ class FakeFilesystem:
 
     @staticmethod
     def _can_read(target, owner_can_read):
-        if target.st_uid == USER_ID:
+        if target.st_uid == helpers.USER_ID:
             if owner_can_read or target.st_mode & 0o400:
                 return True
-        if target.st_gid == GROUP_ID:
+        if target.st_gid == helpers.GROUP_ID:
             if target.st_mode & 0o040:
                 return True
         return target.st_mode & 0o004
@@ -2407,7 +1668,7 @@ class FakeFilesystem:
                 if not self.is_windows_fs and isinstance(parent_obj, FakeFile):
                     self.raise_os_error(errno.ENOTDIR, path_str)
                 self.raise_os_error(errno.ENOENT, path_str)
-            if not parent_obj.st_mode & PERM_READ:
+            if not parent_obj.st_mode & helpers.PERM_READ:
                 self.raise_os_error(errno.EACCES, parent_directory)
             return (
                 parent_obj.get_entry(to_string(child_name))
@@ -2691,7 +1952,7 @@ class FakeFilesystem:
         return path_str.replace(os_sep, fake_sep)  # type: ignore[return-value]
 
     def create_dir(
-        self, directory_path: AnyPath, perm_bits: int = PERM_DEF
+        self, directory_path: AnyPath, perm_bits: int = helpers.PERM_DEF
     ) -> FakeDirectory:
         """Create `directory_path`, and all the parent directories.
 
@@ -2744,7 +2005,7 @@ class FakeFilesystem:
     def create_file(
         self,
         file_path: AnyPath,
-        st_mode: int = S_IFREG | PERM_DEF_FILE,
+        st_mode: int = S_IFREG | helpers.PERM_DEF_FILE,
         contents: AnyString = "",
         st_size: Optional[int] = None,
         create_missing_dirs: bool = True,
@@ -2982,7 +2243,7 @@ class FakeFilesystem:
     def create_file_internally(
         self,
         file_path: AnyPath,
-        st_mode: int = S_IFREG | PERM_DEF_FILE,
+        st_mode: int = S_IFREG | helpers.PERM_DEF_FILE,
         contents: AnyString = "",
         st_size: Optional[int] = None,
         create_missing_dirs: bool = True,
@@ -3120,7 +2381,7 @@ class FakeFilesystem:
             link_path = self.resolve_path(link_path)
         return self.create_file_internally(
             link_path,
-            st_mode=S_IFLNK | PERM_DEF,
+            st_mode=S_IFLNK | helpers.PERM_DEF,
             contents=link_target_path,
             create_missing_dirs=create_missing_dirs,
         )
@@ -3264,7 +2525,7 @@ class FakeFilesystem:
         assert link_obj.contents
         return link_obj.contents
 
-    def makedir(self, dir_path: AnyPath, mode: int = PERM_DEF) -> None:
+    def makedir(self, dir_path: AnyPath, mode: int = helpers.PERM_DEF) -> None:
         """Create a leaf Fake directory.
 
         Args:
@@ -3318,7 +2579,7 @@ class FakeFilesystem:
         return path
 
     def makedirs(
-        self, dir_name: AnyStr, mode: int = PERM_DEF, exist_ok: bool = False
+        self, dir_name: AnyStr, mode: int = helpers.PERM_DEF, exist_ok: bool = False
     ) -> None:
         """Create a leaf Fake directory and create any non-existent
         parent dirs.
@@ -3587,2701 +2848,6 @@ class FakeFilesystem:
         self._add_open_file(StandardStreamWrapper(sys.stdin))
         self._add_open_file(StandardStreamWrapper(sys.stdout))
         self._add_open_file(StandardStreamWrapper(sys.stderr))
-
-
-class FakePathModule:
-    """Faked os.path module replacement.
-
-    FakePathModule should *only* be instantiated by FakeOsModule.  See the
-    FakeOsModule docstring for details.
-    """
-
-    _OS_PATH_COPY: Any = _copy_module(os.path)
-
-    devnull: ClassVar[str] = ""
-    sep: ClassVar[str] = ""
-    altsep: ClassVar[Optional[str]] = None
-    linesep: ClassVar[str] = ""
-    pathsep: ClassVar[str] = ""
-
-    @staticmethod
-    def dir() -> List[str]:
-        """Return the list of patched function names. Used for patching
-        functions imported from the module.
-        """
-        dir_list = [
-            "abspath",
-            "dirname",
-            "exists",
-            "expanduser",
-            "getatime",
-            "getctime",
-            "getmtime",
-            "getsize",
-            "isabs",
-            "isdir",
-            "isfile",
-            "islink",
-            "ismount",
-            "join",
-            "lexists",
-            "normcase",
-            "normpath",
-            "realpath",
-            "relpath",
-            "split",
-            "splitdrive",
-            "samefile",
-        ]
-        if sys.version_info >= (3, 12):
-            dir_list.append("isjunction")
-        return dir_list
-
-    def __init__(self, filesystem: FakeFilesystem, os_module: "FakeOsModule"):
-        """Init.
-
-        Args:
-            filesystem: FakeFilesystem used to provide file system information
-        """
-        self.filesystem = filesystem
-        self._os_path = self._OS_PATH_COPY
-        self._os_path.os = self.os = os_module  # type: ignore[attr-defined]
-        self.reset(filesystem)
-
-    @classmethod
-    def reset(cls, filesystem: FakeFilesystem) -> None:
-        cls.sep = filesystem.path_separator
-        cls.altsep = filesystem.alternative_path_separator
-        cls.linesep = filesystem.line_separator()
-        cls.devnull = "nul" if filesystem.is_windows_fs else "/dev/null"
-        cls.pathsep = ";" if filesystem.is_windows_fs else ":"
-
-    def exists(self, path: AnyStr) -> bool:
-        """Determine whether the file object exists within the fake filesystem.
-
-        Args:
-            path: The path to the file object.
-
-        Returns:
-            (bool) `True` if the file exists.
-        """
-        return self.filesystem.exists(path)
-
-    def lexists(self, path: AnyStr) -> bool:
-        """Test whether a path exists.  Returns True for broken symbolic links.
-
-        Args:
-          path:  path to the symlink object.
-
-        Returns:
-          bool (if file exists).
-        """
-        return self.filesystem.exists(path, check_link=True)
-
-    def getsize(self, path: AnyStr):
-        """Return the file object size in bytes.
-
-        Args:
-          path:  path to the file object.
-
-        Returns:
-          file size in bytes.
-        """
-        file_obj = self.filesystem.resolve(path)
-        if (
-            self.filesystem.ends_with_path_separator(path)
-            and S_IFMT(file_obj.st_mode) != S_IFDIR
-        ):
-            error_nr = errno.EINVAL if self.filesystem.is_windows_fs else errno.ENOTDIR
-            self.filesystem.raise_os_error(error_nr, path)
-        return file_obj.st_size
-
-    def isabs(self, path: AnyStr) -> bool:
-        """Return True if path is an absolute pathname."""
-        if self.filesystem.is_windows_fs:
-            path = self.splitdrive(path)[1]
-        path = make_string_path(path)
-        return self.filesystem.starts_with_sep(path)
-
-    def isdir(self, path: AnyStr) -> bool:
-        """Determine if path identifies a directory."""
-        return self.filesystem.isdir(path)
-
-    def isfile(self, path: AnyStr) -> bool:
-        """Determine if path identifies a regular file."""
-        return self.filesystem.isfile(path)
-
-    def islink(self, path: AnyStr) -> bool:
-        """Determine if path identifies a symbolic link.
-
-        Args:
-            path: Path to filesystem object.
-
-        Returns:
-            `True` if path points to a symbolic link.
-
-        Raises:
-            TypeError: if path is None.
-        """
-        return self.filesystem.islink(path)
-
-    if sys.version_info >= (3, 12):
-
-        def isjunction(self, path: AnyStr) -> bool:
-            """Returns False. Junctions are never faked."""
-            return self.filesystem.isjunction(path)
-
-    def getmtime(self, path: AnyStr) -> float:
-        """Returns the modification time of the fake file.
-
-        Args:
-            path: the path to fake file.
-
-        Returns:
-            (int, float) the modification time of the fake file
-                         in number of seconds since the epoch.
-
-        Raises:
-            OSError: if the file does not exist.
-        """
-        try:
-            file_obj = self.filesystem.resolve(path)
-            return file_obj.st_mtime
-        except OSError:
-            self.filesystem.raise_os_error(errno.ENOENT, winerror=3)
-
-    def getatime(self, path: AnyStr) -> float:
-        """Returns the last access time of the fake file.
-
-        Note: Access time is not set automatically in fake filesystem
-            on access.
-
-        Args:
-            path: the path to fake file.
-
-        Returns:
-            (int, float) the access time of the fake file in number of seconds
-                since the epoch.
-
-        Raises:
-            OSError: if the file does not exist.
-        """
-        try:
-            file_obj = self.filesystem.resolve(path)
-        except OSError:
-            self.filesystem.raise_os_error(errno.ENOENT)
-        return file_obj.st_atime
-
-    def getctime(self, path: AnyStr) -> float:
-        """Returns the creation time of the fake file.
-
-        Args:
-            path: the path to fake file.
-
-        Returns:
-            (int, float) the creation time of the fake file in number of
-                seconds since the epoch.
-
-        Raises:
-            OSError: if the file does not exist.
-        """
-        try:
-            file_obj = self.filesystem.resolve(path)
-        except OSError:
-            self.filesystem.raise_os_error(errno.ENOENT)
-        return file_obj.st_ctime
-
-    def abspath(self, path: AnyStr) -> AnyStr:
-        """Return the absolute version of a path."""
-
-        def getcwd():
-            """Return the current working directory."""
-            # pylint: disable=undefined-variable
-            if isinstance(path, bytes):
-                return self.os.getcwdb()
-            else:
-                return self.os.getcwd()
-
-        path = make_string_path(path)
-        if not self.isabs(path):
-            path = self.join(getcwd(), path)
-        elif self.filesystem.is_windows_fs and self.filesystem.starts_with_sep(path):
-            cwd = getcwd()
-            if self.filesystem.starts_with_drive_letter(cwd):
-                path = self.join(cwd[:2], path)
-        return self.normpath(path)
-
-    def join(self, *p: AnyStr) -> AnyStr:
-        """Return the completed path with a separator of the parts."""
-        return self.filesystem.joinpaths(*p)
-
-    def split(self, path: AnyStr) -> Tuple[AnyStr, AnyStr]:
-        """Split the path into the directory and the filename of the path."""
-        return self.filesystem.splitpath(path)
-
-    def splitdrive(self, path: AnyStr) -> Tuple[AnyStr, AnyStr]:
-        """Split the path into the drive part and the rest of the path, if
-        supported."""
-        return self.filesystem.splitdrive(path)
-
-    def normpath(self, path: AnyStr) -> AnyStr:
-        """Normalize path, eliminating double slashes, etc."""
-        return self.filesystem.normpath(path)
-
-    def normcase(self, path: AnyStr) -> AnyStr:
-        """Convert to lower case under windows, replaces additional path
-        separator."""
-        path = self.filesystem.normcase(path)
-        if self.filesystem.is_windows_fs:
-            path = path.lower()
-        return path
-
-    def relpath(self, path: AnyStr, start: Optional[AnyStr] = None) -> AnyStr:
-        """We mostly rely on the native implementation and adapt the
-        path separator."""
-        if not path:
-            raise ValueError("no path specified")
-        path = make_string_path(path)
-        path = self.filesystem.replace_windows_root(path)
-        sep = matching_string(path, self.filesystem.path_separator)
-        if start is not None:
-            start = make_string_path(start)
-        else:
-            start = matching_string(path, self.filesystem.cwd)
-        start = self.filesystem.replace_windows_root(start)
-        system_sep = matching_string(path, self._os_path.sep)
-        if self.filesystem.alternative_path_separator is not None:
-            altsep = matching_string(path, self.filesystem.alternative_path_separator)
-            path = path.replace(altsep, system_sep)
-            start = start.replace(altsep, system_sep)
-        path = path.replace(sep, system_sep)
-        start = start.replace(sep, system_sep)
-        path = self._os_path.relpath(path, start)
-        return path.replace(system_sep, sep)
-
-    def realpath(self, filename: AnyStr, strict: Optional[bool] = None) -> AnyStr:
-        """Return the canonical path of the specified filename, eliminating any
-        symbolic links encountered in the path.
-        """
-        if strict is not None and sys.version_info < (3, 10):
-            raise TypeError("realpath() got an unexpected " "keyword argument 'strict'")
-        if strict:
-            # raises in strict mode if the file does not exist
-            self.filesystem.resolve(filename)
-        if self.filesystem.is_windows_fs:
-            return self.abspath(filename)
-        filename = make_string_path(filename)
-        path, ok = self._join_real_path(filename[:0], filename, {})
-        path = self.abspath(path)
-        return path
-
-    def samefile(self, path1: AnyStr, path2: AnyStr) -> bool:
-        """Return whether path1 and path2 point to the same file.
-
-        Args:
-            path1: first file path or path object (Python >=3.6)
-            path2: second file path or path object (Python >=3.6)
-
-        Raises:
-            OSError: if one of the paths does not point to an existing
-                file system object.
-        """
-        stat1 = self.filesystem.stat(path1)
-        stat2 = self.filesystem.stat(path2)
-        return stat1.st_ino == stat2.st_ino and stat1.st_dev == stat2.st_dev
-
-    @overload
-    def _join_real_path(
-        self, path: str, rest: str, seen: Dict[str, Optional[str]]
-    ) -> Tuple[str, bool]:
-        ...
-
-    @overload
-    def _join_real_path(
-        self, path: bytes, rest: bytes, seen: Dict[bytes, Optional[bytes]]
-    ) -> Tuple[bytes, bool]:
-        ...
-
-    def _join_real_path(
-        self, path: AnyStr, rest: AnyStr, seen: Dict[AnyStr, Optional[AnyStr]]
-    ) -> Tuple[AnyStr, bool]:
-        """Join two paths, normalizing and eliminating any symbolic links
-        encountered in the second path.
-        Taken from Python source and adapted.
-        """
-        curdir = matching_string(path, ".")
-        pardir = matching_string(path, "..")
-
-        sep = self.filesystem.get_path_separator(path)
-        if self.isabs(rest):
-            rest = rest[1:]
-            path = sep
-
-        while rest:
-            name, _, rest = rest.partition(sep)
-            if not name or name == curdir:
-                # current dir
-                continue
-            if name == pardir:
-                # parent dir
-                if path:
-                    path, name = self.filesystem.splitpath(path)
-                    if name == pardir:
-                        path = self.filesystem.joinpaths(path, pardir, pardir)
-                else:
-                    path = pardir
-                continue
-            newpath = self.filesystem.joinpaths(path, name)
-            if not self.filesystem.islink(newpath):
-                path = newpath
-                continue
-            # Resolve the symbolic link
-            if newpath in seen:
-                # Already seen this path
-                seen_path = seen[newpath]
-                if seen_path is not None:
-                    # use cached value
-                    path = seen_path
-                    continue
-                # The symlink is not resolved, so we must have a symlink loop.
-                # Return already resolved part + rest of the path unchanged.
-                return self.filesystem.joinpaths(newpath, rest), False
-            seen[newpath] = None  # not resolved symlink
-            path, ok = self._join_real_path(
-                path,
-                matching_string(path, self.filesystem.readlink(newpath)),
-                seen,
-            )
-            if not ok:
-                return self.filesystem.joinpaths(path, rest), False
-            seen[newpath] = path  # resolved symlink
-        return path, True
-
-    def dirname(self, path: AnyStr) -> AnyStr:
-        """Returns the first part of the result of `split()`."""
-        return self.split(path)[0]
-
-    def expanduser(self, path: AnyStr) -> AnyStr:
-        """Return the argument with an initial component of ~ or ~user
-        replaced by that user's home directory.
-        """
-        path = self._os_path.expanduser(path)
-        return path.replace(
-            matching_string(path, self._os_path.sep),
-            matching_string(path, self.sep),
-        )
-
-    def ismount(self, path: AnyStr) -> bool:
-        """Return true if the given path is a mount point.
-
-        Args:
-            path: Path to filesystem object to be checked
-
-        Returns:
-            `True` if path is a mount point added to the fake file system.
-            Under Windows also returns True for drive and UNC roots
-            (independent of their existence).
-        """
-        if not path:
-            return False
-        path_str = to_string(make_string_path(path))
-        normed_path = self.filesystem.absnormpath(path_str)
-        sep = self.filesystem.path_separator
-        if self.filesystem.is_windows_fs:
-            path_seps: Union[Tuple[str, Optional[str]], Tuple[str]]
-            if self.filesystem.alternative_path_separator is not None:
-                path_seps = (sep, self.filesystem.alternative_path_separator)
-            else:
-                path_seps = (sep,)
-            drive, rest = self.filesystem.splitdrive(normed_path)
-            if drive and drive[:1] in path_seps:
-                return (not rest) or (rest in path_seps)
-            if rest in path_seps:
-                return True
-        for mount_point in self.filesystem.mount_points:
-            if to_string(normed_path).rstrip(sep) == to_string(mount_point).rstrip(sep):
-                return True
-        return False
-
-    def __getattr__(self, name: str) -> Any:
-        """Forwards any non-faked calls to the real os.path."""
-        return getattr(self._os_path, name)
-
-
-class FakeOsModule:
-    """Uses FakeFilesystem to provide a fake os module replacement.
-
-    Do not create os.path separately from os, as there is a necessary circular
-    dependency between os and os.path to replicate the behavior of the standard
-    Python modules.  What you want to do is to just let FakeOsModule take care
-    of `os.path` setup itself.
-
-    # You always want to do this.
-    filesystem = fake_filesystem.FakeFilesystem()
-    my_os_module = fake_filesystem.FakeOsModule(filesystem)
-    """
-
-    use_original = False
-
-    @staticmethod
-    def dir() -> List[str]:
-        """Return the list of patched function names. Used for patching
-        functions imported from the module.
-        """
-        _dir = [
-            "access",
-            "chdir",
-            "chmod",
-            "chown",
-            "close",
-            "fstat",
-            "fsync",
-            "getcwd",
-            "lchmod",
-            "link",
-            "listdir",
-            "lstat",
-            "makedirs",
-            "mkdir",
-            "mknod",
-            "open",
-            "read",
-            "readlink",
-            "remove",
-            "removedirs",
-            "rename",
-            "rmdir",
-            "stat",
-            "symlink",
-            "umask",
-            "unlink",
-            "utime",
-            "walk",
-            "write",
-            "getcwdb",
-            "replace",
-        ]
-        if sys.platform.startswith("linux"):
-            _dir += [
-                "fdatasync",
-                "getxattr",
-                "listxattr",
-                "removexattr",
-                "setxattr",
-            ]
-        if use_scandir:
-            _dir += ["scandir"]
-        return _dir
-
-    def __init__(self, filesystem: FakeFilesystem):
-        """Also exposes self.path (to fake os.path).
-
-        Args:
-            filesystem: FakeFilesystem used to provide file system information
-        """
-        self.filesystem = filesystem
-        self.os_module: Any = os
-        self.path = FakePathModule(self.filesystem, self)
-
-    @property
-    def devnull(self) -> str:
-        return self.path.devnull
-
-    @property
-    def sep(self) -> str:
-        return self.path.sep
-
-    @property
-    def altsep(self) -> Optional[str]:
-        return self.path.altsep
-
-    @property
-    def linesep(self) -> str:
-        return self.path.linesep
-
-    @property
-    def pathsep(self) -> str:
-        return self.path.pathsep
-
-    def fdopen(self, fd: int, *args: Any, **kwargs: Any) -> AnyFileWrapper:
-        """Redirector to open() builtin function.
-
-        Args:
-            fd: The file descriptor of the file to open.
-            *args: Pass through args.
-            **kwargs: Pass through kwargs.
-
-        Returns:
-            File object corresponding to file_des.
-
-        Raises:
-            TypeError: if file descriptor is not an integer.
-        """
-        if not is_int_type(fd):
-            raise TypeError("an integer is required")
-        return FakeFileOpen(self.filesystem)(fd, *args, **kwargs)
-
-    def _umask(self) -> int:
-        """Return the current umask."""
-        if self.filesystem.is_windows_fs:
-            # windows always returns 0 - it has no real notion of umask
-            return 0
-        if sys.platform == "win32":
-            # if we are testing Unix under Windows we assume a default mask
-            return 0o002
-        else:
-            # under Unix, we return the real umask;
-            # as there is no pure getter for umask, so we have to first
-            # set a mode to get the previous one and then re-set that
-            mask = os.umask(0)
-            os.umask(mask)
-            return mask
-
-    def open(
-        self,
-        path: AnyStr,
-        flags: int,
-        mode: Optional[int] = None,
-        *,
-        dir_fd: Optional[int] = None
-    ) -> int:
-        """Return the file descriptor for a FakeFile.
-
-        Args:
-            path: the path to the file
-            flags: low-level bits to indicate io operation
-            mode: bits to define default permissions
-                Note: only basic modes are supported, OS-specific modes are
-                ignored
-            dir_fd: If not `None`, the file descriptor of a directory,
-                with `file_path` being relative to this directory.
-
-        Returns:
-            A file descriptor.
-
-        Raises:
-            OSError: if the path cannot be found
-            ValueError: if invalid mode is given
-            NotImplementedError: if `os.O_EXCL` is used without `os.O_CREAT`
-        """
-        path = self._path_with_dir_fd(path, self.open, dir_fd)
-        if mode is None:
-            if self.filesystem.is_windows_fs:
-                mode = 0o666
-            else:
-                mode = 0o777 & ~self._umask()
-
-        has_tmpfile_flag = (
-            hasattr(os, "O_TMPFILE") and flags & os.O_TMPFILE == os.O_TMPFILE
-        )
-        open_modes = _OpenModes(
-            must_exist=not flags & os.O_CREAT and not has_tmpfile_flag,
-            can_read=not flags & os.O_WRONLY,
-            can_write=flags & (os.O_RDWR | os.O_WRONLY) != 0,
-            truncate=flags & os.O_TRUNC != 0,
-            append=flags & os.O_APPEND != 0,
-            must_not_exist=flags & os.O_EXCL != 0,
-        )
-        if open_modes.must_not_exist and open_modes.must_exist:
-            raise NotImplementedError("O_EXCL without O_CREAT mode is not supported")
-        if has_tmpfile_flag:
-            # this is a workaround for tempfiles that do not have a filename
-            # as we do not support this directly, we just add a unique filename
-            # and set the file to delete on close
-            path = self.filesystem.joinpaths(
-                path, matching_string(path, str(uuid.uuid4()))
-            )
-
-        if not self.filesystem.is_windows_fs and self.filesystem.exists(path):
-            # handle opening directory - only allowed under Posix
-            # with read-only mode
-            obj = self.filesystem.resolve(path)
-            if isinstance(obj, FakeDirectory):
-                if (
-                    not open_modes.must_exist and not self.filesystem.is_macos
-                ) or open_modes.can_write:
-                    self.filesystem.raise_os_error(errno.EISDIR, path)
-                dir_wrapper = FakeDirWrapper(obj, path, self.filesystem)
-                file_des = self.filesystem._add_open_file(dir_wrapper)
-                dir_wrapper.filedes = file_des
-                return file_des
-
-        # low level open is always binary
-        str_flags = "b"
-        delete_on_close = has_tmpfile_flag
-        if hasattr(os, "O_TEMPORARY"):
-            delete_on_close = flags & os.O_TEMPORARY == os.O_TEMPORARY
-        fake_file = FakeFileOpen(
-            self.filesystem, delete_on_close=delete_on_close, raw_io=True
-        )(path, str_flags, open_modes=open_modes)
-        assert not isinstance(fake_file, StandardStreamWrapper)
-        if fake_file.file_object != self.filesystem.dev_null:
-            self.chmod(path, mode)
-        return fake_file.fileno()
-
-    def close(self, fd: int) -> None:
-        """Close a file descriptor.
-
-        Args:
-            fd: An integer file descriptor for the file object requested.
-
-        Raises:
-            OSError: bad file descriptor.
-            TypeError: if file descriptor is not an integer.
-        """
-        file_handle = self.filesystem.get_open_file(fd)
-        file_handle.close()
-
-    def read(self, fd: int, n: int) -> bytes:
-        """Read number of bytes from a file descriptor, returns bytes read.
-
-        Args:
-            fd: An integer file descriptor for the file object requested.
-            n: Number of bytes to read from file.
-
-        Returns:
-            Bytes read from file.
-
-        Raises:
-            OSError: bad file descriptor.
-            TypeError: if file descriptor is not an integer.
-        """
-        file_handle = self.filesystem.get_open_file(fd)
-        if isinstance(file_handle, FakeFileWrapper):
-            file_handle.raw_io = True
-        if isinstance(file_handle, FakeDirWrapper):
-            self.filesystem.raise_os_error(errno.EBADF, file_handle.file_path)
-        return file_handle.read(n)
-
-    def write(self, fd: int, contents: bytes) -> int:
-        """Write string to file descriptor, returns number of bytes written.
-
-        Args:
-            fd: An integer file descriptor for the file object requested.
-            contents: String of bytes to write to file.
-
-        Returns:
-            Number of bytes written.
-
-        Raises:
-            OSError: bad file descriptor.
-            TypeError: if file descriptor is not an integer.
-        """
-        file_handle = cast(FakeFileWrapper, self.filesystem.get_open_file(fd))
-        if isinstance(file_handle, FakeDirWrapper):
-            self.filesystem.raise_os_error(errno.EBADF, file_handle.file_path)
-
-        if isinstance(file_handle, FakePipeWrapper):
-            return file_handle.write(contents)
-
-        file_handle.raw_io = True
-        file_handle._sync_io()
-        file_handle.update_flush_pos()
-        file_handle.write(contents)
-        file_handle.flush()
-        return len(contents)
-
-    def pipe(self) -> Tuple[int, int]:
-        read_fd, write_fd = os.pipe()
-        read_wrapper = FakePipeWrapper(self.filesystem, read_fd, False)
-        file_des = self.filesystem._add_open_file(read_wrapper)
-        read_wrapper.filedes = file_des
-        write_wrapper = FakePipeWrapper(self.filesystem, write_fd, True)
-        file_des = self.filesystem._add_open_file(write_wrapper)
-        write_wrapper.filedes = file_des
-        return read_wrapper.filedes, write_wrapper.filedes
-
-    def fstat(self, fd: int) -> FakeStatResult:
-        """Return the os.stat-like tuple for the FakeFile object of file_des.
-
-        Args:
-            fd: The file descriptor of filesystem object to retrieve.
-
-        Returns:
-            The FakeStatResult object corresponding to entry_path.
-
-        Raises:
-            OSError: if the filesystem object doesn't exist.
-        """
-        # stat should return the tuple representing return value of os.stat
-        file_object = self.filesystem.get_open_file(fd).get_object()
-        assert isinstance(file_object, FakeFile)
-        return file_object.stat_result.copy()
-
-    def umask(self, mask: int) -> int:
-        """Change the current umask.
-
-        Args:
-            mask: (int) The new umask value.
-
-        Returns:
-            The old umask.
-
-        Raises:
-            TypeError: if new_mask is of an invalid type.
-        """
-        if not is_int_type(mask):
-            raise TypeError("an integer is required")
-        old_umask = self.filesystem.umask
-        self.filesystem.umask = mask
-        return old_umask
-
-    def chdir(self, path: AnyStr) -> None:
-        """Change current working directory to target directory.
-
-        Args:
-            path: The path to new current working directory.
-
-        Raises:
-            OSError: if user lacks permission to enter the argument directory
-                or if the target is not a directory.
-        """
-        try:
-            path = self.filesystem.resolve_path(path, allow_fd=True)
-        except OSError as exc:
-            if self.filesystem.is_macos and exc.errno == errno.EBADF:
-                raise OSError(errno.ENOTDIR, "Not a directory: " + str(path))
-            raise
-        self.filesystem.confirmdir(path)
-        directory = self.filesystem.resolve(path)
-        # A full implementation would check permissions all the way
-        # up the tree.
-        if not is_root() and not directory.st_mode | PERM_EXE:
-            self.filesystem.raise_os_error(errno.EACCES, directory.name)
-        self.filesystem.cwd = path  # type: ignore[assignment]
-
-    def getcwd(self) -> str:
-        """Return current working directory."""
-        return to_string(self.filesystem.cwd)
-
-    def getcwdb(self) -> bytes:
-        """Return current working directory as bytes."""
-        return to_bytes(self.filesystem.cwd)
-
-    def listdir(self, path: AnyStr) -> List[AnyStr]:
-        """Return a list of file names in target_directory.
-
-        Args:
-            path: Path to the target directory within the fake
-                filesystem.
-
-        Returns:
-            A list of file names within the target directory in arbitrary
-                order.
-
-        Raises:
-          OSError:  if the target is not a directory.
-        """
-        return self.filesystem.listdir(path)
-
-    XATTR_CREATE = 1
-    XATTR_REPLACE = 2
-
-    def getxattr(
-        self, path: AnyStr, attribute: AnyString, *, follow_symlinks: bool = True
-    ) -> Optional[bytes]:
-        """Return the value of the given extended filesystem attribute for
-        `path`.
-
-        Args:
-            path: File path, file descriptor or path-like object (for
-                Python >= 3.6).
-            attribute: (str or bytes) The attribute name.
-            follow_symlinks: (bool) If True (the default), symlinks in the
-                path are traversed.
-
-        Returns:
-            The contents of the extended attribute as bytes or None if
-            the attribute does not exist.
-
-        Raises:
-            OSError: if the path does not exist.
-        """
-        if not self.filesystem.is_linux:
-            raise AttributeError("module 'os' has no attribute 'getxattr'")
-
-        if isinstance(attribute, bytes):
-            attribute = attribute.decode(sys.getfilesystemencoding())
-        file_obj = self.filesystem.resolve(path, follow_symlinks, allow_fd=True)
-        return file_obj.xattr.get(attribute)
-
-    def listxattr(
-        self, path: Optional[AnyStr] = None, *, follow_symlinks: bool = True
-    ) -> List[str]:
-        """Return a list of the extended filesystem attributes on `path`.
-
-        Args:
-            path: File path, file descriptor or path-like object (for
-                Python >= 3.6). If None, the current directory is used.
-            follow_symlinks: (bool) If True (the default), symlinks in the
-                path are traversed.
-
-        Returns:
-            A list of all attribute names for the given path as str.
-
-        Raises:
-            OSError: if the path does not exist.
-        """
-        if not self.filesystem.is_linux:
-            raise AttributeError("module 'os' has no attribute 'listxattr'")
-
-        path_str = self.filesystem.cwd if path is None else path
-        file_obj = self.filesystem.resolve(
-            cast(AnyStr, path_str), follow_symlinks, allow_fd=True
-        )
-        return list(file_obj.xattr.keys())
-
-    def removexattr(
-        self, path: AnyStr, attribute: AnyString, *, follow_symlinks: bool = True
-    ) -> None:
-        """Removes the extended filesystem attribute attribute from `path`.
-
-        Args:
-            path: File path, file descriptor or path-like object (for
-                Python >= 3.6).
-            attribute: (str or bytes) The attribute name.
-            follow_symlinks: (bool) If True (the default), symlinks in the
-                path are traversed.
-
-        Raises:
-            OSError: if the path does not exist.
-        """
-        if not self.filesystem.is_linux:
-            raise AttributeError("module 'os' has no attribute 'removexattr'")
-
-        if isinstance(attribute, bytes):
-            attribute = attribute.decode(sys.getfilesystemencoding())
-        file_obj = self.filesystem.resolve(path, follow_symlinks, allow_fd=True)
-        if attribute in file_obj.xattr:
-            del file_obj.xattr[attribute]
-
-    def setxattr(
-        self,
-        path: AnyStr,
-        attribute: AnyString,
-        value: bytes,
-        flags: int = 0,
-        *,
-        follow_symlinks: bool = True
-    ) -> None:
-        """Sets the value of the given extended filesystem attribute for
-        `path`.
-
-        Args:
-            path: File path, file descriptor or path-like object (for
-                Python >= 3.6).
-            attribute: The attribute name (str or bytes).
-            value: (byte-like) The value to be set.
-            follow_symlinks: (bool) If True (the default), symlinks in the
-                path are traversed.
-
-        Raises:
-            OSError: if the path does not exist.
-            TypeError: if `value` is not a byte-like object.
-        """
-        if not self.filesystem.is_linux:
-            raise AttributeError("module 'os' has no attribute 'setxattr'")
-
-        if isinstance(attribute, bytes):
-            attribute = attribute.decode(sys.getfilesystemencoding())
-        if not is_byte_string(value):
-            raise TypeError("a bytes-like object is required")
-        file_obj = self.filesystem.resolve(path, follow_symlinks, allow_fd=True)
-        exists = attribute in file_obj.xattr
-        if exists and flags == self.XATTR_CREATE:
-            self.filesystem.raise_os_error(errno.ENODATA, file_obj.path)
-        if not exists and flags == self.XATTR_REPLACE:
-            self.filesystem.raise_os_error(errno.EEXIST, file_obj.path)
-        file_obj.xattr[attribute] = value
-
-    def scandir(self, path: str = ".") -> ScanDirIter:
-        """Return an iterator of DirEntry objects corresponding to the
-        entries in the directory given by path.
-
-        Args:
-            path: Path to the target directory within the fake filesystem.
-
-        Returns:
-            An iterator to an unsorted list of os.DirEntry objects for
-            each entry in path.
-
-        Raises:
-            OSError: if the target is not a directory.
-        """
-        return scandir(self.filesystem, path)
-
-    def walk(
-        self,
-        top: AnyStr,
-        topdown: bool = True,
-        onerror: Optional[bool] = None,
-        followlinks: bool = False,
-    ):
-        """Perform an os.walk operation over the fake filesystem.
-
-        Args:
-            top: The root directory from which to begin walk.
-            topdown: Determines whether to return the tuples with the root as
-                the first entry (`True`) or as the last, after all the child
-                directory tuples (`False`).
-          onerror: If not `None`, function which will be called to handle the
-                `os.error` instance provided when `os.listdir()` fails.
-          followlinks: If `True`, symbolic links are followed.
-
-        Yields:
-            (path, directories, nondirectories) for top and each of its
-            subdirectories.  See the documentation for the builtin os module
-            for further details.
-        """
-        return walk(self.filesystem, top, topdown, onerror, followlinks)
-
-    def readlink(self, path: AnyStr, dir_fd: Optional[int] = None) -> str:
-        """Read the target of a symlink.
-
-        Args:
-            path:  Symlink to read the target of.
-            dir_fd: If not `None`, the file descriptor of a directory,
-                with `path` being relative to this directory.
-
-        Returns:
-            the string representing the path to which the symbolic link points.
-
-        Raises:
-            TypeError: if `path` is None
-            OSError: (with errno=ENOENT) if path is not a valid path, or
-                     (with errno=EINVAL) if path is valid, but is not a symlink
-        """
-        path = self._path_with_dir_fd(path, self.readlink, dir_fd)
-        return self.filesystem.readlink(path)
-
-    def stat(
-        self,
-        path: AnyStr,
-        *,
-        dir_fd: Optional[int] = None,
-        follow_symlinks: bool = True
-    ) -> FakeStatResult:
-        """Return the os.stat-like tuple for the FakeFile object of entry_path.
-
-        Args:
-            path:  path to filesystem object to retrieve.
-            dir_fd: (int) If not `None`, the file descriptor of a directory,
-                with `entry_path` being relative to this directory.
-            follow_symlinks: (bool) If `False` and `entry_path` points to a
-                symlink, the link itself is changed instead of the linked
-                object.
-
-        Returns:
-            The FakeStatResult object corresponding to entry_path.
-
-        Raises:
-            OSError: if the filesystem object doesn't exist.
-        """
-        path = self._path_with_dir_fd(path, self.stat, dir_fd)
-        return self.filesystem.stat(path, follow_symlinks)
-
-    def lstat(self, path: AnyStr, *, dir_fd: Optional[int] = None) -> FakeStatResult:
-        """Return the os.stat-like tuple for entry_path,
-        not following symlinks.
-
-        Args:
-            path:  path to filesystem object to retrieve.
-            dir_fd: If not `None`, the file descriptor of a directory, with
-                `path` being relative to this directory.
-
-        Returns:
-            the FakeStatResult object corresponding to `path`.
-
-        Raises:
-            OSError: if the filesystem object doesn't exist.
-        """
-        # stat should return the tuple representing return value of os.stat
-        path = self._path_with_dir_fd(path, self.lstat, dir_fd)
-        return self.filesystem.stat(path, follow_symlinks=False)
-
-    def remove(self, path: AnyStr, dir_fd: Optional[int] = None) -> None:
-        """Remove the FakeFile object at the specified file path.
-
-        Args:
-            path: Path to file to be removed.
-            dir_fd: If not `None`, the file descriptor of a directory,
-                with `path` being relative to this directory.
-
-        Raises:
-            OSError: if path points to a directory.
-            OSError: if path does not exist.
-            OSError: if removal failed.
-        """
-        path = self._path_with_dir_fd(path, self.remove, dir_fd)
-        self.filesystem.remove(path)
-
-    def unlink(self, path: AnyStr, *, dir_fd: Optional[int] = None) -> None:
-        """Remove the FakeFile object at the specified file path.
-
-        Args:
-            path: Path to file to be removed.
-            dir_fd: If not `None`, the file descriptor of a directory,
-                with `path` being relative to this directory.
-
-        Raises:
-            OSError: if path points to a directory.
-            OSError: if path does not exist.
-            OSError: if removal failed.
-        """
-        path = self._path_with_dir_fd(path, self.unlink, dir_fd)
-        self.filesystem.remove(path)
-
-    def rename(
-        self,
-        src: AnyStr,
-        dst: AnyStr,
-        *,
-        src_dir_fd: Optional[int] = None,
-        dst_dir_fd: Optional[int] = None
-    ) -> None:
-        """Rename a FakeFile object at old_file_path to new_file_path,
-        preserving all properties.
-        Also replaces existing new_file_path object, if one existed
-        (Unix only).
-
-        Args:
-            src: Path to filesystem object to rename.
-            dst: Path to where the filesystem object will live
-                after this call.
-            src_dir_fd: If not `None`, the file descriptor of a directory,
-                with `src` being relative to this directory.
-            dst_dir_fd: If not `None`, the file descriptor of a directory,
-                with `dst` being relative to this directory.
-
-        Raises:
-            OSError: if old_file_path does not exist.
-            OSError: if new_file_path is an existing directory.
-            OSError: if new_file_path is an existing file (Windows only)
-            OSError: if new_file_path is an existing file and could not
-                be removed (Unix)
-            OSError: if `dirname(new_file)` does not exist
-            OSError: if the file would be moved to another filesystem
-                (e.g. mount point)
-        """
-        src = self._path_with_dir_fd(src, self.rename, src_dir_fd)
-        dst = self._path_with_dir_fd(dst, self.rename, dst_dir_fd)
-        self.filesystem.rename(src, dst)
-
-    def renames(self, old: AnyStr, new: AnyStr):
-        """Fakes `os.renames`, documentation taken from there.
-
-        Super-rename; create directories as necessary and delete any left
-        empty.  Works like rename, except creation of any intermediate
-        directories needed to make the new pathname good is attempted
-        first.  After the rename, directories corresponding to rightmost
-        path segments of the old name will be pruned until either the
-        whole path is consumed or a nonempty directory is found.
-
-        Note: this function can fail with the new directory structure made
-        if you lack permissions needed to unlink the leaf directory or
-        file.
-
-        """
-        head, tail = self.filesystem.splitpath(new)
-        if head and tail and not self.filesystem.exists(head):
-            self.makedirs(head)
-        self.rename(old, new)
-        head, tail = self.filesystem.splitpath(old)
-        if head and tail:
-            try:
-                self.removedirs(head)
-            except OSError:
-                pass
-
-    def replace(
-        self,
-        src: AnyStr,
-        dst: AnyStr,
-        *,
-        src_dir_fd: Optional[int] = None,
-        dst_dir_fd: Optional[int] = None
-    ) -> None:
-        """Renames a FakeFile object at old_file_path to new_file_path,
-        preserving all properties.
-        Also replaces existing new_file_path object, if one existed.
-
-        Arg
-            src: Path to filesystem object to rename.
-            dst: Path to where the filesystem object will live
-                after this call.
-            src_dir_fd: If not `None`, the file descriptor of a directory,
-                with `src` being relative to this directory.
-            dst_dir_fd: If not `None`, the file descriptor of a directory,
-                with `dst` being relative to this directory.
-
-        Raises:
-            OSError: if old_file_path does not exist.
-            OSError: if new_file_path is an existing directory.
-            OSError: if new_file_path is an existing file and could
-                not be removed
-            OSError: if `dirname(new_file)` does not exist
-            OSError: if the file would be moved to another filesystem
-                (e.g. mount point)
-        """
-        src = self._path_with_dir_fd(src, self.rename, src_dir_fd)
-        dst = self._path_with_dir_fd(dst, self.rename, dst_dir_fd)
-        self.filesystem.rename(src, dst, force_replace=True)
-
-    def rmdir(self, path: AnyStr, *, dir_fd: Optional[int] = None) -> None:
-        """Remove a leaf Fake directory.
-
-        Args:
-            path: (str) Name of directory to remove.
-            dir_fd: If not `None`, the file descriptor of a directory,
-                with `path` being relative to this directory.
-
-        Raises:
-            OSError: if `path` does not exist or is not a directory,
-            or as per FakeFilesystem.remove_object. Cannot remove '.'.
-        """
-        path = self._path_with_dir_fd(path, self.rmdir, dir_fd)
-        self.filesystem.rmdir(path)
-
-    def removedirs(self, name: AnyStr) -> None:
-        """Remove a leaf fake directory and all empty intermediate ones.
-
-        Args:
-            name: the directory to be removed.
-
-        Raises:
-            OSError: if target_directory does not exist or is not a directory.
-            OSError: if target_directory is not empty.
-        """
-        name = self.filesystem.absnormpath(name)
-        directory = self.filesystem.confirmdir(name)
-        if directory.entries:
-            self.filesystem.raise_os_error(errno.ENOTEMPTY, self.path.basename(name))
-        else:
-            self.rmdir(name)
-        head, tail = self.path.split(name)
-        if not tail:
-            head, tail = self.path.split(head)
-        while head and tail:
-            head_dir = self.filesystem.confirmdir(head)
-            if head_dir.entries:
-                break
-            # only the top-level dir may not be a symlink
-            self.filesystem.rmdir(head, allow_symlink=True)
-            head, tail = self.path.split(head)
-
-    def mkdir(
-        self, path: AnyStr, mode: int = PERM_DEF, *, dir_fd: Optional[int] = None
-    ) -> None:
-        """Create a leaf Fake directory.
-
-        Args:
-            path: (str) Name of directory to create.
-                Relative paths are assumed to be relative to '/'.
-            mode: (int) Mode to create directory with.  This argument defaults
-                to 0o777.  The umask is applied to this mode.
-            dir_fd: If not `None`, the file descriptor of a directory,
-                with `path` being relative to this directory.
-
-        Raises:
-            OSError: if the directory name is invalid or parent directory is
-                read only or as per FakeFilesystem.add_object.
-        """
-        path = self._path_with_dir_fd(path, self.mkdir, dir_fd)
-        try:
-            self.filesystem.makedir(path, mode)
-        except OSError as e:
-            if e.errno == errno.EACCES:
-                self.filesystem.raise_os_error(e.errno, path)
-            raise
-
-    def makedirs(
-        self, name: AnyStr, mode: int = PERM_DEF, exist_ok: Optional[bool] = None
-    ) -> None:
-        """Create a leaf Fake directory + create any non-existent parent dirs.
-
-        Args:
-            name: (str) Name of directory to create.
-            mode: (int) Mode to create directory (and any necessary parent
-                directories) with. This argument defaults to 0o777.
-                The umask is applied to this mode.
-            exist_ok: (boolean) If exist_ok is False (the default), an OSError
-                is raised if the target directory already exists.
-
-        Raises:
-            OSError: if the directory already exists and exist_ok=False, or as
-                per :py:meth:`FakeFilesystem.create_dir`.
-        """
-        if exist_ok is None:
-            exist_ok = False
-        self.filesystem.makedirs(name, mode, exist_ok)
-
-    def _path_with_dir_fd(
-        self, path: AnyStr, fct: Callable, dir_fd: Optional[int]
-    ) -> AnyStr:
-        """Return the path considering dir_fd. Raise on invalid parameters."""
-        try:
-            path = make_string_path(path)
-        except TypeError:
-            # the error is handled later
-            path = path
-        if dir_fd is not None:
-            # check if fd is supported for the built-in real function
-            real_fct = getattr(os, fct.__name__)
-            if real_fct not in self.supports_dir_fd:
-                raise NotImplementedError("dir_fd unavailable on this platform")
-            if isinstance(path, int):
-                raise ValueError(
-                    "%s: Can't specify dir_fd without "
-                    "matching path_str" % fct.__name__
-                )
-            if not self.path.isabs(path):
-                open_file = self.filesystem.get_open_file(dir_fd)
-                return self.path.join(  # type: ignore[type-var, return-value]
-                    cast(FakeFile, open_file.get_object()).path, path
-                )
-        return path
-
-    def truncate(self, path: AnyStr, length: int) -> None:
-        """Truncate the file corresponding to path, so that it is
-         length bytes in size. If length is larger than the current size,
-         the file is filled up with zero bytes.
-
-        Args:
-            path: (str or int) Path to the file, or an integer file
-                descriptor for the file object.
-            length: (int) Length of the file after truncating it.
-
-        Raises:
-            OSError: if the file does not exist or the file descriptor is
-                invalid.
-        """
-        file_object = self.filesystem.resolve(path, allow_fd=True)
-        file_object.size = length
-
-    def ftruncate(self, fd: int, length: int) -> None:
-        """Truncate the file corresponding to fd, so that it is
-         length bytes in size. If length is larger than the current size,
-         the file is filled up with zero bytes.
-
-        Args:
-            fd: (int) File descriptor for the file object.
-            length: (int) Maximum length of the file after truncating it.
-
-        Raises:
-            OSError: if the file descriptor is invalid
-        """
-        file_object = self.filesystem.get_open_file(fd).get_object()
-        if isinstance(file_object, FakeFileWrapper):
-            file_object.size = length
-        else:
-            raise OSError(errno.EBADF, "Invalid file descriptor")
-
-    def access(
-        self,
-        path: AnyStr,
-        mode: int,
-        *,
-        dir_fd: Optional[int] = None,
-        effective_ids: bool = False,
-        follow_symlinks: bool = True
-    ) -> bool:
-        """Check if a file exists and has the specified permissions.
-
-        Args:
-            path: (str) Path to the file.
-            mode: (int) Permissions represented as a bitwise-OR combination of
-                os.F_OK, os.R_OK, os.W_OK, and os.X_OK.
-            dir_fd: If not `None`, the file descriptor of a directory, with
-                `path` being relative to this directory.
-            effective_ids: (bool) Unused. Only here to match the signature.
-            follow_symlinks: (bool) If `False` and `path` points to a symlink,
-                the link itself is queried instead of the linked object.
-
-        Returns:
-            bool, `True` if file is accessible, `False` otherwise.
-        """
-        if effective_ids and self.filesystem.is_windows_fs:
-            raise NotImplementedError(
-                "access: effective_ids unavailable on this platform"
-            )
-        path = self._path_with_dir_fd(path, self.access, dir_fd)
-        try:
-            stat_result = self.stat(path, follow_symlinks=follow_symlinks)
-        except OSError as os_error:
-            if os_error.errno == errno.ENOENT:
-                return False
-            raise
-        if is_root():
-            mode &= ~os.W_OK
-        return (mode & ((stat_result.st_mode >> 6) & 7)) == mode
-
-    def chmod(
-        self,
-        path: AnyStr,
-        mode: int,
-        *,
-        dir_fd: Optional[int] = None,
-        follow_symlinks: bool = True
-    ) -> None:
-        """Change the permissions of a file as encoded in integer mode.
-
-        Args:
-            path: (str) Path to the file.
-            mode: (int) Permissions.
-            dir_fd: If not `None`, the file descriptor of a directory, with
-                `path` being relative to this directory.
-            follow_symlinks: (bool) If `False` and `path` points to a symlink,
-                the link itself is queried instead of the linked object.
-        """
-        if not follow_symlinks and (
-            os.chmod not in os.supports_follow_symlinks or IS_PYPY
-        ):
-            raise NotImplementedError(
-                "`follow_symlinks` for chmod() is not available " "on this system"
-            )
-        path = self._path_with_dir_fd(path, self.chmod, dir_fd)
-        self.filesystem.chmod(path, mode, follow_symlinks)
-
-    def lchmod(self, path: AnyStr, mode: int) -> None:
-        """Change the permissions of a file as encoded in integer mode.
-        If the file is a link, the permissions of the link are changed.
-
-        Args:
-          path: (str) Path to the file.
-          mode: (int) Permissions.
-        """
-        if self.filesystem.is_windows_fs:
-            raise NameError("name 'lchmod' is not defined")
-        self.filesystem.chmod(path, mode, follow_symlinks=False)
-
-    def utime(
-        self,
-        path: AnyStr,
-        times: Optional[Tuple[Union[int, float], Union[int, float]]] = None,
-        ns: Optional[Tuple[int, int]] = None,
-        dir_fd: Optional[int] = None,
-        follow_symlinks: bool = True,
-    ) -> None:
-        """Change the access and modified times of a file.
-
-        Args:
-            path: (str) Path to the file.
-            times: 2-tuple of int or float numbers, of the form (atime, mtime)
-                which is used to set the access and modified times in seconds.
-                If None, both times are set to the current time.
-            ns: 2-tuple of int numbers, of the form (atime, mtime)  which is
-                used to set the access and modified times in nanoseconds.
-                If None, both times are set to the current time.
-            dir_fd: If not `None`, the file descriptor of a directory,
-                with `path` being relative to this directory.
-            follow_symlinks: (bool) If `False` and `path` points to a symlink,
-                the link itself is queried instead of the linked object.
-
-            Raises:
-                TypeError: If anything other than the expected types is
-                    specified in the passed `times` or `ns` tuple,
-                    or if the tuple length is not equal to 2.
-                ValueError: If both times and ns are specified.
-        """
-        path = self._path_with_dir_fd(path, self.utime, dir_fd)
-        self.filesystem.utime(path, times=times, ns=ns, follow_symlinks=follow_symlinks)
-
-    def chown(
-        self,
-        path: AnyStr,
-        uid: int,
-        gid: int,
-        *,
-        dir_fd: Optional[int] = None,
-        follow_symlinks: bool = True
-    ) -> None:
-        """Set ownership of a faked file.
-
-        Args:
-            path: (str) Path to the file or directory.
-            uid: (int) Numeric uid to set the file or directory to.
-            gid: (int) Numeric gid to set the file or directory to.
-            dir_fd: (int) If not `None`, the file descriptor of a directory,
-                with `path` being relative to this directory.
-            follow_symlinks: (bool) If `False` and path points to a symlink,
-                the link itself is changed instead of the linked object.
-
-        Raises:
-            OSError: if path does not exist.
-
-        `None` is also allowed for `uid` and `gid`.  This permits `os.rename`
-        to use `os.chown` even when the source file `uid` and `gid` are
-        `None` (unset).
-        """
-        path = self._path_with_dir_fd(path, self.chown, dir_fd)
-        file_object = self.filesystem.resolve(path, follow_symlinks, allow_fd=True)
-        if not isinstance(uid, int) or not isinstance(gid, int):
-            raise TypeError("An integer is required")
-        if uid != -1:
-            file_object.st_uid = uid
-        if gid != -1:
-            file_object.st_gid = gid
-
-    def mknod(
-        self,
-        path: AnyStr,
-        mode: Optional[int] = None,
-        device: int = 0,
-        *,
-        dir_fd: Optional[int] = None
-    ) -> None:
-        """Create a filesystem node named 'filename'.
-
-        Does not support device special files or named pipes as the real os
-        module does.
-
-        Args:
-            path: (str) Name of the file to create
-            mode: (int) Permissions to use and type of file to be created.
-                Default permissions are 0o666.  Only the stat.S_IFREG file type
-                is supported by the fake implementation.  The umask is applied
-                to this mode.
-            device: not supported in fake implementation
-            dir_fd: If not `None`, the file descriptor of a directory,
-                with `path` being relative to this directory.
-
-        Raises:
-          OSError: if called with unsupported options or the file can not be
-          created.
-        """
-        if self.filesystem.is_windows_fs:
-            raise AttributeError("module 'os' has no attribute 'mknode'")
-        if mode is None:
-            # note that a default value of 0o600 without a device type is
-            # documented - this is not how it seems to work
-            mode = S_IFREG | 0o600
-        if device or not mode & S_IFREG and not is_root():
-            self.filesystem.raise_os_error(errno.EPERM)
-
-        path = self._path_with_dir_fd(path, self.mknod, dir_fd)
-        head, tail = self.path.split(path)
-        if not tail:
-            if self.filesystem.exists(head, check_link=True):
-                self.filesystem.raise_os_error(errno.EEXIST, path)
-            self.filesystem.raise_os_error(errno.ENOENT, path)
-        if tail in (matching_string(tail, "."), matching_string(tail, "..")):
-            self.filesystem.raise_os_error(errno.ENOENT, path)
-        if self.filesystem.exists(path, check_link=True):
-            self.filesystem.raise_os_error(errno.EEXIST, path)
-        self.filesystem.add_object(
-            head,
-            FakeFile(tail, mode & ~self.filesystem.umask, filesystem=self.filesystem),
-        )
-
-    def symlink(
-        self,
-        src: AnyStr,
-        dst: AnyStr,
-        target_is_directory: bool = False,
-        *,
-        dir_fd: Optional[int] = None
-    ) -> None:
-        """Creates the specified symlink, pointed at the specified link target.
-
-        Args:
-            src: The target of the symlink.
-            dst: Path to the symlink to create.
-            target_is_directory: Currently ignored.
-            dir_fd: If not `None`, the file descriptor of a directory,
-                with `src` being relative to this directory.
-
-        Raises:
-            OSError:  if the file already exists.
-        """
-        src = self._path_with_dir_fd(src, self.symlink, dir_fd)
-        self.filesystem.create_symlink(dst, src, create_missing_dirs=False)
-
-    def link(
-        self,
-        src: AnyStr,
-        dst: AnyStr,
-        *,
-        src_dir_fd: Optional[int] = None,
-        dst_dir_fd: Optional[int] = None
-    ) -> None:
-        """Create a hard link at new_path, pointing at old_path.
-
-        Args:
-            src: An existing path to the target file.
-            dst: The destination path to create a new link at.
-            src_dir_fd: If not `None`, the file descriptor of a directory,
-                with `src` being relative to this directory.
-            dst_dir_fd: If not `None`, the file descriptor of a directory,
-                with `dst` being relative to this directory.
-
-        Raises:
-            OSError:  if something already exists at new_path.
-            OSError:  if the parent directory doesn't exist.
-        """
-        src = self._path_with_dir_fd(src, self.link, src_dir_fd)
-        dst = self._path_with_dir_fd(dst, self.link, dst_dir_fd)
-        self.filesystem.link(src, dst)
-
-    def fsync(self, fd: int) -> None:
-        """Perform fsync for a fake file (in other words, do nothing).
-
-        Args:
-            fd: The file descriptor of the open file.
-
-        Raises:
-            OSError: file_des is an invalid file descriptor.
-            TypeError: file_des is not an integer.
-        """
-        # Throw an error if file_des isn't valid
-        if 0 <= fd < NR_STD_STREAMS:
-            self.filesystem.raise_os_error(errno.EINVAL)
-        file_object = cast(FakeFileWrapper, self.filesystem.get_open_file(fd))
-        if self.filesystem.is_windows_fs:
-            if not hasattr(file_object, "allow_update") or not file_object.allow_update:
-                self.filesystem.raise_os_error(errno.EBADF, file_object.file_path)
-
-    def fdatasync(self, fd: int) -> None:
-        """Perform fdatasync for a fake file (in other words, do nothing).
-
-        Args:
-            fd: The file descriptor of the open file.
-
-        Raises:
-            OSError: `fd` is an invalid file descriptor.
-            TypeError: `fd` is not an integer.
-        """
-        if self.filesystem.is_windows_fs or self.filesystem.is_macos:
-            raise AttributeError("module 'os' has no attribute 'fdatasync'")
-        # Throw an error if file_des isn't valid
-        if 0 <= fd < NR_STD_STREAMS:
-            self.filesystem.raise_os_error(errno.EINVAL)
-        self.filesystem.get_open_file(fd)
-
-    def sendfile(self, fd_out: int, fd_in: int, offset: int, count: int) -> int:
-        """Copy count bytes from file descriptor fd_in to file descriptor
-        fd_out starting at offset.
-
-        Args:
-            fd_out: The file descriptor of the destination file.
-            fd_in: The file descriptor of the source file.
-            offset: The offset in bytes where to start the copy in the
-                source file. If `None` (Linux only), copying is started at
-                the current position, and the position is updated.
-            count: The number of bytes to copy. If 0, all remaining bytes
-                are copied (MacOs only).
-
-        Raises:
-            OSError: If `fd_in` or `fd_out` is an invalid file descriptor.
-            TypeError: If `fd_in` or `fd_out` is not an integer.
-            TypeError: If `offset` is None under MacOs.
-        """
-        if self.filesystem.is_windows_fs:
-            raise AttributeError("module 'os' has no attribute 'sendfile'")
-        if 0 <= fd_in < NR_STD_STREAMS:
-            self.filesystem.raise_os_error(errno.EINVAL)
-        if 0 <= fd_out < NR_STD_STREAMS:
-            self.filesystem.raise_os_error(errno.EINVAL)
-        source = cast(FakeFileWrapper, self.filesystem.get_open_file(fd_in))
-        dest = cast(FakeFileWrapper, self.filesystem.get_open_file(fd_out))
-        if self.filesystem.is_macos:
-            if dest.get_object().stat_result.st_mode & 0o777000 != S_IFSOCK:
-                raise OSError("Socket operation on non-socket")
-        if offset is None:
-            if self.filesystem.is_macos:
-                raise TypeError("None is not a valid offset")
-            contents = source.read(count)
-        else:
-            position = source.tell()
-            source.seek(offset)
-            if count == 0 and self.filesystem.is_macos:
-                contents = source.read()
-            else:
-                contents = source.read(count)
-            source.seek(position)
-        if contents:
-            written = dest.write(contents)
-            dest.flush()
-            return written
-        return 0
-
-    def __getattr__(self, name: str) -> Any:
-        """Forwards any unfaked calls to the standard os module."""
-        return getattr(self.os_module, name)
-
-
-if sys.version_info > (3, 10):
-
-    def handle_original_call(f: Callable) -> Callable:
-        """Decorator used for real pathlib Path methods to ensure that
-        real os functions instead of faked ones are used.
-        Applied to all non-private methods of `FakeOsModule`."""
-
-        @functools.wraps(f)
-        def wrapped(*args, **kwargs):
-            if not f.__name__.startswith("_") and FakeOsModule.use_original:
-                # remove the `self` argument for FakeOsModule methods
-                if args and isinstance(args[0], FakeOsModule):
-                    args = args[1:]
-                return getattr(os, f.__name__)(*args, **kwargs)
-            return f(*args, **kwargs)
-
-        return wrapped
-
-    for name, fn in inspect.getmembers(FakeOsModule, inspect.isfunction):
-        setattr(FakeOsModule, name, handle_original_call(fn))
-
-
-@contextmanager
-def use_original_os():
-    """Temporarily use original os functions instead of faked ones.
-    Used to ensure that skipped modules do not use faked calls.
-    """
-    try:
-        FakeOsModule.use_original = True
-        yield
-    finally:
-        FakeOsModule.use_original = False
-
-
-class FakeIoModule:
-    """Uses FakeFilesystem to provide a fake io module replacement.
-
-    You need a fake_filesystem to use this:
-    filesystem = fake_filesystem.FakeFilesystem()
-    my_io_module = fake_filesystem.FakeIoModule(filesystem)
-    """
-
-    @staticmethod
-    def dir() -> List[str]:
-        """Return the list of patched function names. Used for patching
-        functions imported from the module.
-        """
-        _dir = ["open"]
-        if sys.version_info >= (3, 8):
-            _dir.append("open_code")
-        return _dir
-
-    def __init__(self, filesystem: FakeFilesystem):
-        """
-        Args:
-            filesystem: FakeFilesystem used to provide file system information.
-        """
-        self.filesystem = filesystem
-        self.skip_names: List[str] = []
-        self._io_module = io
-
-    def open(
-        self,
-        file: Union[AnyStr, int],
-        mode: str = "r",
-        buffering: int = -1,
-        encoding: Optional[str] = None,
-        errors: Optional[str] = None,
-        newline: Optional[str] = None,
-        closefd: bool = True,
-        opener: Optional[Callable] = None,
-    ) -> Union[AnyFileWrapper, IO[Any]]:
-        """Redirect the call to FakeFileOpen.
-        See FakeFileOpen.call() for description.
-        """
-        # workaround for built-in open called from skipped modules (see #552)
-        # as open is not imported explicitly, we cannot patch it for
-        # specific modules; instead we check if the caller is a skipped
-        # module (should work in most cases)
-        stack = traceback.extract_stack(limit=2)
-        module_name = os.path.splitext(stack[0].filename)[0]
-        module_name = module_name.replace(os.sep, ".")
-        if any(
-            [
-                module_name == sn or module_name.endswith("." + sn)
-                for sn in self.skip_names
-            ]
-        ):
-            return io.open(
-                file,
-                mode,
-                buffering,
-                encoding,
-                errors,
-                newline,
-                closefd,
-                opener,
-            )
-        fake_open = FakeFileOpen(self.filesystem)
-        return fake_open(
-            file, mode, buffering, encoding, errors, newline, closefd, opener
-        )
-
-    if sys.version_info >= (3, 8):
-
-        def open_code(self, path):
-            """Redirect the call to open. Note that the behavior of the real
-            function may be overridden by an earlier call to the
-            PyFile_SetOpenCodeHook(). This behavior is not reproduced here.
-            """
-            if not isinstance(path, str):
-                raise TypeError("open_code() argument 'path' must be str, not int")
-            patch_mode = self.filesystem.patch_open_code
-            if (
-                patch_mode == PatchMode.AUTO
-                and self.filesystem.exists(path)
-                or patch_mode == PatchMode.ON
-            ):
-                return self.open(path, mode="rb")
-            # mostly this is used for compiled code -
-            # don't patch these, as the files are probably in the real fs
-            return self._io_module.open_code(path)
-
-    def __getattr__(self, name):
-        """Forwards any unfaked calls to the standard io module."""
-        return getattr(self._io_module, name)
-
-
-if sys.platform != "win32":
-    import fcntl
-
-    class FakeFcntlModule:
-        """Replaces the fcntl module. Only valid under Linux/MacOS,
-        currently just mocks the functionality away.
-        """
-
-        @staticmethod
-        def dir() -> List[str]:
-            """Return the list of patched function names. Used for patching
-            functions imported from the module.
-            """
-            return ["fcntl", "ioctl", "flock", "lockf"]
-
-        def __init__(self, filesystem: FakeFilesystem):
-            """
-            Args:
-                filesystem: FakeFilesystem used to provide file system
-                    information (currently not used).
-            """
-            self.filesystem = filesystem
-            self._fcntl_module = fcntl
-
-        def fcntl(self, fd: int, cmd: int, arg: int = 0) -> Union[int, bytes]:
-            return 0 if isinstance(arg, int) else arg
-
-        def ioctl(
-            self, fd: int, request: int, arg: int = 0, mutate_flag: bool = True
-        ) -> Union[int, bytes]:
-            return 0 if isinstance(arg, int) else arg
-
-        def flock(self, fd: int, operation: int) -> None:
-            pass
-
-        def lockf(
-            self, fd: int, cmd: int, len: int = 0, start: int = 0, whence=0
-        ) -> Any:
-            pass
-
-        def __getattr__(self, name):
-            """Forwards any unfaked calls to the standard fcntl module."""
-            return getattr(self._fcntl_module, name)
-
-
-class FakeFileWrapper:
-    """Wrapper for a stream object for use by a FakeFile object.
-
-    If the wrapper has any data written to it, it will propagate to
-    the FakeFile object on close() or flush().
-    """
-
-    def __init__(
-        self,
-        file_object: FakeFile,
-        file_path: AnyStr,
-        update: bool,
-        read: bool,
-        append: bool,
-        delete_on_close: bool,
-        filesystem: FakeFilesystem,
-        newline: Optional[str],
-        binary: bool,
-        closefd: bool,
-        encoding: Optional[str],
-        errors: Optional[str],
-        buffering: int,
-        raw_io: bool,
-        is_stream: bool = False,
-    ):
-        self.file_object = file_object
-        self.file_path = file_path  # type: ignore[var-annotated]
-        self._append = append
-        self._read = read
-        self.allow_update = update
-        self._closefd = closefd
-        self._file_epoch = file_object.epoch
-        self.raw_io = raw_io
-        self._binary = binary
-        self.is_stream = is_stream
-        self._changed = False
-        self._buffer_size = buffering
-        if self._buffer_size == 0 and not binary:
-            raise ValueError("can't have unbuffered text I/O")
-        # buffer_size is ignored in text mode
-        elif self._buffer_size == -1 or not binary:
-            self._buffer_size = io.DEFAULT_BUFFER_SIZE
-        self._use_line_buffer = not binary and buffering == 1
-
-        contents = file_object.byte_contents
-        self._encoding = encoding or locale.getpreferredencoding(False)
-        errors = errors or "strict"
-        self._io: Union[BinaryBufferIO, TextBufferIO] = (
-            BinaryBufferIO(contents)
-            if binary
-            else TextBufferIO(
-                contents, encoding=encoding, newline=newline, errors=errors
-            )
-        )
-        self._read_whence = 0
-        self._read_seek = 0
-        self._flush_pos = 0
-        if contents:
-            self._flush_pos = len(contents)
-            if update:
-                if not append:
-                    self._io.seek(0)
-                else:
-                    self._io.seek(self._flush_pos)
-                    self._read_seek = self._io.tell()
-
-        if delete_on_close:
-            assert filesystem, "delete_on_close=True requires filesystem"
-        self._filesystem = filesystem
-        self.delete_on_close = delete_on_close
-        # override, don't modify FakeFile.name, as FakeFilesystem expects
-        # it to be the file name only, no directories.
-        self.name = file_object.opened_as
-        self.filedes: Optional[int] = None
-
-    def __enter__(self) -> "FakeFileWrapper":
-        """To support usage of this fake file with the 'with' statement."""
-        return self
-
-    def __exit__(
-        self,
-        exc_type: Optional[Type[BaseException]],
-        exc_val: Optional[BaseException],
-        exc_tb: Optional[TracebackType],
-    ) -> None:
-        """To support usage of this fake file with the 'with' statement."""
-        self.close()
-
-    def _raise(self, message: str) -> NoReturn:
-        if self.raw_io:
-            self._filesystem.raise_os_error(errno.EBADF, self.file_path)
-        raise io.UnsupportedOperation(message)
-
-    def get_object(self) -> FakeFile:
-        """Return the FakeFile object that is wrapped
-        by the current instance.
-        """
-        return self.file_object
-
-    def fileno(self) -> int:
-        """Return the file descriptor of the file object."""
-        if self.filedes is not None:
-            return self.filedes
-        raise OSError(errno.EBADF, "Invalid file descriptor")
-
-    def close(self) -> None:
-        """Close the file."""
-        # ignore closing a closed file
-        if not self._is_open():
-            return
-
-        # for raw io, all writes are flushed immediately
-        if self.allow_update and not self.raw_io:
-            self.flush()
-            if self._filesystem.is_windows_fs and self._changed:
-                self.file_object.st_mtime = now()
-
-        assert self.filedes is not None
-        if self._closefd:
-            self._filesystem._close_open_file(self.filedes)
-        else:
-            open_files = self._filesystem.open_files[self.filedes]
-            assert open_files is not None
-            open_files.remove(self)
-        if self.delete_on_close:
-            self._filesystem.remove_object(
-                self.get_object().path  # type: ignore[arg-type]
-            )
-
-    @property
-    def closed(self) -> bool:
-        """Simulate the `closed` attribute on file."""
-        return not self._is_open()
-
-    def _try_flush(self, old_pos: int) -> None:
-        """Try to flush and reset the position if it fails."""
-        flush_pos = self._flush_pos
-        try:
-            self.flush()
-        except OSError:
-            # write failed - reset to previous position
-            self._io.seek(old_pos)
-            self._io.truncate()
-            self._flush_pos = flush_pos
-            raise
-
-    def flush(self) -> None:
-        """Flush file contents to 'disk'."""
-        self._check_open_file()
-        if self.allow_update and not self.is_stream:
-            contents = self._io.getvalue()
-            if self._append:
-                self._sync_io()
-                old_contents = self.file_object.byte_contents
-                assert old_contents is not None
-                contents = old_contents + contents[self._flush_pos :]
-                self._set_stream_contents(contents)
-            else:
-                self._io.flush()
-            changed = self.file_object.set_contents(contents, self._encoding)
-            self.update_flush_pos()
-            if changed:
-                if self._filesystem.is_windows_fs:
-                    self._changed = True
-                else:
-                    current_time = now()
-                    self.file_object.st_ctime = current_time
-                    self.file_object.st_mtime = current_time
-            self._file_epoch = self.file_object.epoch
-
-            if not self.is_stream:
-                self._flush_related_files()
-
-    def update_flush_pos(self) -> None:
-        self._flush_pos = self._io.tell()
-
-    def _flush_related_files(self) -> None:
-        for open_files in self._filesystem.open_files[3:]:
-            if open_files is not None:
-                for open_file in open_files:
-                    if (
-                        open_file is not self
-                        and isinstance(open_file, FakeFileWrapper)
-                        and self.file_object == open_file.file_object
-                        and not open_file._append
-                    ):
-                        open_file._sync_io()
-
-    def seek(self, offset: int, whence: int = 0) -> None:
-        """Move read/write pointer in 'file'."""
-        self._check_open_file()
-        if not self._append:
-            self._io.seek(offset, whence)
-        else:
-            self._read_seek = offset
-            self._read_whence = whence
-        if not self.is_stream:
-            self.flush()
-
-    def tell(self) -> int:
-        """Return the file's current position.
-
-        Returns:
-          int, file's current position in bytes.
-        """
-        self._check_open_file()
-        if not self.is_stream:
-            self.flush()
-
-        if not self._append:
-            return self._io.tell()
-        if self._read_whence:
-            write_seek = self._io.tell()
-            self._io.seek(self._read_seek, self._read_whence)
-            self._read_seek = self._io.tell()
-            self._read_whence = 0
-            self._io.seek(write_seek)
-        return self._read_seek
-
-    def _sync_io(self) -> None:
-        """Update the stream with changes to the file object contents."""
-        if self._file_epoch == self.file_object.epoch:
-            return
-
-        contents = self.file_object.byte_contents
-        assert contents is not None
-        self._set_stream_contents(contents)
-        self._file_epoch = self.file_object.epoch
-
-    def _set_stream_contents(self, contents: bytes) -> None:
-        whence = self._io.tell()
-        self._io.seek(0)
-        self._io.truncate()
-        self._io.putvalue(contents)
-        if not self._append:
-            self._io.seek(whence)
-
-    def _read_wrappers(self, name: str) -> Callable:
-        """Wrap a stream attribute in a read wrapper.
-
-        Returns a read_wrapper which tracks our own read pointer since the
-        stream object has no concept of a different read and write pointer.
-
-        Args:
-            name: The name of the attribute to wrap. Should be a read call.
-
-        Returns:
-            The read_wrapper function.
-        """
-        io_attr = getattr(self._io, name)
-
-        def read_wrapper(*args, **kwargs):
-            """Wrap all read calls to the stream object.
-
-            We do this to track the read pointer separate from the write
-            pointer.  Anything that wants to read from the stream object
-            while we're in append mode goes through this.
-
-            Args:
-                *args: pass through args
-                **kwargs: pass through kwargs
-            Returns:
-                Wrapped stream object method
-            """
-            self._io.seek(self._read_seek, self._read_whence)
-            ret_value = io_attr(*args, **kwargs)
-            self._read_seek = self._io.tell()
-            self._read_whence = 0
-            self._io.seek(0, 2)
-            return ret_value
-
-        return read_wrapper
-
-    def _other_wrapper(self, name: str) -> Callable:
-        """Wrap a stream attribute in an other_wrapper.
-
-        Args:
-          name: the name of the stream attribute to wrap.
-
-        Returns:
-          other_wrapper which is described below.
-        """
-        io_attr = getattr(self._io, name)
-
-        def other_wrapper(*args, **kwargs):
-            """Wrap all other calls to the stream Object.
-
-            We do this to track changes to the write pointer.  Anything that
-            moves the write pointer in a file open for appending should move
-            the read pointer as well.
-
-            Args:
-                *args: Pass through args.
-                **kwargs: Pass through kwargs.
-
-            Returns:
-                Wrapped stream object method.
-            """
-            write_seek = self._io.tell()
-            ret_value = io_attr(*args, **kwargs)
-            if write_seek != self._io.tell():
-                self._read_seek = self._io.tell()
-                self._read_whence = 0
-
-            return ret_value
-
-        return other_wrapper
-
-    def _write_wrapper(self, name: str) -> Callable:
-        """Wrap a stream attribute in a write_wrapper.
-
-        Args:
-          name: the name of the stream attribute to wrap.
-
-        Returns:
-          write_wrapper which is described below.
-        """
-        io_attr = getattr(self._io, name)
-
-        def write_wrapper(*args, **kwargs):
-            """Wrap all other calls to the stream Object.
-
-            We do this to track changes to the write pointer.  Anything that
-            moves the write pointer in a file open for appending should move
-            the read pointer as well.
-
-            Args:
-                *args: Pass through args.
-                **kwargs: Pass through kwargs.
-
-            Returns:
-                Wrapped stream object method.
-            """
-            old_pos = self._io.tell()
-            ret_value = io_attr(*args, **kwargs)
-            new_pos = self._io.tell()
-
-            # if the buffer size is exceeded, we flush
-            use_line_buf = self._use_line_buffer and "\n" in args[0]
-            if new_pos - self._flush_pos > self._buffer_size or use_line_buf:
-                flush_all = new_pos - old_pos > self._buffer_size or use_line_buf
-                # if the current write does not exceed the buffer size,
-                # we revert to the previous position and flush that,
-                # otherwise we flush all
-                if not flush_all:
-                    self._io.seek(old_pos)
-                    self._io.truncate()
-                self._try_flush(old_pos)
-                if not flush_all:
-                    ret_value = io_attr(*args, **kwargs)
-            if self._append:
-                self._read_seek = self._io.tell()
-                self._read_whence = 0
-            return ret_value
-
-        return write_wrapper
-
-    def _adapt_size_for_related_files(self, size: int) -> None:
-        for open_files in self._filesystem.open_files[3:]:
-            if open_files is not None:
-                for open_file in open_files:
-                    if (
-                        open_file is not self
-                        and isinstance(open_file, FakeFileWrapper)
-                        and self.file_object == open_file.file_object
-                        and cast(FakeFileWrapper, open_file)._append
-                    ):
-                        open_file._read_seek += size
-
-    def _truncate_wrapper(self) -> Callable:
-        """Wrap truncate() to allow flush after truncate.
-
-        Returns:
-            Wrapper which is described below.
-        """
-        io_attr = self._io.truncate
-
-        def truncate_wrapper(*args, **kwargs):
-            """Wrap truncate call to call flush after truncate."""
-            if self._append:
-                self._io.seek(self._read_seek, self._read_whence)
-            size = io_attr(*args, **kwargs)
-            self.flush()
-            if not self.is_stream:
-                self.file_object.size = size
-                buffer_size = len(self._io.getvalue())
-                if buffer_size < size:
-                    self._io.seek(buffer_size)
-                    self._io.putvalue(b"\0" * (size - buffer_size))
-                    self.file_object.set_contents(self._io.getvalue(), self._encoding)
-                    self._flush_pos = size
-                    self._adapt_size_for_related_files(size - buffer_size)
-
-            self.flush()
-            return size
-
-        return truncate_wrapper
-
-    def size(self) -> int:
-        """Return the content size in bytes of the wrapped file."""
-        return self.file_object.st_size
-
-    def __getattr__(self, name: str) -> Any:
-        if self.file_object.is_large_file():
-            raise FakeLargeFileIoException(self.file_path)
-
-        reading = name.startswith("read") or name == "next"
-        truncate = name == "truncate"
-        writing = name.startswith("write") or truncate
-
-        if reading or writing:
-            self._check_open_file()
-        if not self._read and reading:
-            return self._read_error()
-        if not self.allow_update and writing:
-            return self._write_error()
-
-        if reading:
-            self._sync_io()
-            if not self.is_stream:
-                self.flush()
-            if not self._filesystem.is_windows_fs:
-                self.file_object.st_atime = now()
-        if truncate:
-            return self._truncate_wrapper()
-        if self._append:
-            if reading:
-                return self._read_wrappers(name)
-            elif not writing:
-                return self._other_wrapper(name)
-        if writing:
-            return self._write_wrapper(name)
-
-        return getattr(self._io, name)
-
-    def _read_error(self) -> Callable:
-        def read_error(*args, **kwargs):
-            """Throw an error unless the argument is zero."""
-            if args and args[0] == 0:
-                if self._filesystem.is_windows_fs and self.raw_io:
-                    return b"" if self._binary else ""
-            self._raise("File is not open for reading.")
-
-        return read_error
-
-    def _write_error(self) -> Callable:
-        def write_error(*args, **kwargs):
-            """Throw an error."""
-            if self.raw_io:
-                if self._filesystem.is_windows_fs and args and len(args[0]) == 0:
-                    return 0
-            self._raise("File is not open for writing.")
-
-        return write_error
-
-    def _is_open(self) -> bool:
-        if self.filedes is not None and self.filedes < len(self._filesystem.open_files):
-            open_files = self._filesystem.open_files[self.filedes]
-            if open_files is not None and self in open_files:
-                return True
-        return False
-
-    def _check_open_file(self) -> None:
-        if not self.is_stream and not self._is_open():
-            raise ValueError("I/O operation on closed file")
-
-    def __iter__(self) -> Union[Iterator[str], Iterator[bytes]]:
-        if not self._read:
-            self._raise("File is not open for reading")
-        return self._io.__iter__()
-
-    def __next__(self):
-        if not self._read:
-            self._raise("File is not open for reading")
-        return next(self._io)
-
-
-class StandardStreamWrapper:
-    """Wrapper for a system standard stream to be used in open files list."""
-
-    def __init__(self, stream_object: TextIO):
-        self._stream_object = stream_object
-        self.filedes: Optional[int] = None
-
-    def get_object(self) -> TextIO:
-        return self._stream_object
-
-    def fileno(self) -> int:
-        """Return the file descriptor of the wrapped standard stream."""
-        if self.filedes is not None:
-            return self.filedes
-        raise OSError(errno.EBADF, "Invalid file descriptor")
-
-    def read(self, n: int = -1) -> bytes:
-        return cast(bytes, self._stream_object.read())
-
-    def close(self) -> None:
-        """We do not support closing standard streams."""
-
-    def is_stream(self) -> bool:
-        return True
-
-
-class FakeDirWrapper:
-    """Wrapper for a FakeDirectory object to be used in open files list."""
-
-    def __init__(
-        self,
-        file_object: FakeDirectory,
-        file_path: AnyString,
-        filesystem: FakeFilesystem,
-    ):
-        self.file_object = file_object
-        self.file_path = file_path
-        self._filesystem = filesystem
-        self.filedes: Optional[int] = None
-
-    def get_object(self) -> FakeDirectory:
-        """Return the FakeFile object that is wrapped by the current
-        instance."""
-        return self.file_object
-
-    def fileno(self) -> int:
-        """Return the file descriptor of the file object."""
-        if self.filedes is not None:
-            return self.filedes
-        raise OSError(errno.EBADF, "Invalid file descriptor")
-
-    def close(self) -> None:
-        """Close the directory."""
-        assert self.filedes is not None
-        self._filesystem._close_open_file(self.filedes)
-
-
-class FakePipeWrapper:
-    """Wrapper for a read or write descriptor of a real pipe object to be
-    used in open files list.
-    """
-
-    def __init__(
-        self,
-        filesystem: FakeFilesystem,
-        fd: int,
-        can_write: bool,
-        mode: str = "",
-    ):
-        self._filesystem = filesystem
-        self.fd = fd  # the real file descriptor
-        self.can_write = can_write
-        self.file_object = None
-        self.filedes: Optional[int] = None
-        self.real_file = None
-        if mode:
-            self.real_file = open(fd, mode)
-
-    def __enter__(self) -> "FakePipeWrapper":
-        """To support usage of this fake pipe with the 'with' statement."""
-        return self
-
-    def __exit__(
-        self,
-        exc_type: Optional[Type[BaseException]],
-        exc_val: Optional[BaseException],
-        exc_tb: Optional[TracebackType],
-    ) -> None:
-        """To support usage of this fake pipe with the 'with' statement."""
-        self.close()
-
-    def get_object(self) -> None:
-        return self.file_object
-
-    def fileno(self) -> int:
-        """Return the fake file descriptor of the pipe object."""
-        if self.filedes is not None:
-            return self.filedes
-        raise OSError(errno.EBADF, "Invalid file descriptor")
-
-    def read(self, numBytes: int = -1) -> bytes:
-        """Read from the real pipe."""
-        if self.real_file:
-            return self.real_file.read(numBytes)
-        return os.read(self.fd, numBytes)
-
-    def flush(self) -> None:
-        """Flush the real pipe?"""
-
-    def write(self, contents: bytes) -> int:
-        """Write to the real pipe."""
-        if self.real_file:
-            return self.real_file.write(contents)
-        return os.write(self.fd, contents)
-
-    def close(self) -> None:
-        """Close the pipe descriptor."""
-        assert self.filedes is not None
-        open_files = self._filesystem.open_files[self.filedes]
-        assert open_files is not None
-        open_files.remove(self)
-        if self.real_file:
-            self.real_file.close()
-        else:
-            os.close(self.fd)
-
-    def readable(self) -> bool:
-        """The pipe end can either be readable or writable."""
-        return not self.can_write
-
-    def writable(self) -> bool:
-        """The pipe end can either be readable or writable."""
-        return self.can_write
-
-    def seekable(self) -> bool:
-        """A pipe is not seekable."""
-        return False
-
-
-class FakeFileOpen:
-    """Faked `file()` and `open()` function replacements.
-
-    Returns FakeFile objects in a FakeFilesystem in place of the `file()`
-    or `open()` function.
-    """
-
-    __name__ = "FakeFileOpen"
-
-    def __init__(
-        self,
-        filesystem: FakeFilesystem,
-        delete_on_close: bool = False,
-        raw_io: bool = False,
-    ):
-        """
-        Args:
-          filesystem:  FakeFilesystem used to provide file system information
-          delete_on_close:  optional boolean, deletes file on close()
-        """
-        self.filesystem = filesystem
-        self._delete_on_close = delete_on_close
-        self.raw_io = raw_io
-
-    def __call__(self, *args: Any, **kwargs: Any) -> AnyFileWrapper:
-        """Redirects calls to file() or open() to appropriate method."""
-        return self.call(*args, **kwargs)
-
-    def call(
-        self,
-        file_: Union[AnyStr, int],
-        mode: str = "r",
-        buffering: int = -1,
-        encoding: Optional[str] = None,
-        errors: Optional[str] = None,
-        newline: Optional[str] = None,
-        closefd: bool = True,
-        opener: Any = None,
-        open_modes: Optional[_OpenModes] = None,
-    ) -> AnyFileWrapper:
-        """Return a file-like object with the contents of the target
-        file object.
-
-        Args:
-            file_: Path to target file or a file descriptor.
-            mode: Additional file modes (all modes in `open()` are supported).
-            buffering: the buffer size used for writing. Data will only be
-                flushed if buffer size is exceeded. The default (-1) uses a
-                system specific default buffer size. Text line mode (e.g.
-                buffering=1 in text mode) is not supported.
-            encoding: The encoding used to encode unicode strings / decode
-                bytes.
-            errors: (str) Defines how encoding errors are handled.
-            newline: Controls universal newlines, passed to stream object.
-            closefd: If a file descriptor rather than file name is passed,
-                and this is set to `False`, then the file descriptor is kept
-                open when file is closed.
-            opener: an optional function object that will be called with
-                `file_` and the open flags (derived from `mode`) and returns
-                a file descriptor.
-            open_modes: Modes for opening files if called from low-level API.
-
-        Returns:
-            A file-like object containing the contents of the target file.
-
-        Raises:
-            OSError depending on Python version / call mode:
-                - if the target object is a directory
-                - on an invalid path
-                - if the file does not exist when it should
-                - if the file exists but should not
-                - if permission is denied
-            ValueError: for an invalid mode or mode combination
-        """
-        binary = "b" in mode
-
-        if binary and encoding:
-            raise ValueError("binary mode doesn't take an encoding argument")
-
-        newline, open_modes = self._handle_file_mode(mode, newline, open_modes)
-
-        # the pathlib opener is defined in a Path instance that may not be
-        # patched under some circumstances; as it just calls standard open(),
-        # we may ignore it, as it would not change the behavior
-        if opener is not None and opener.__module__ != "pathlib":
-            # opener shall return a file descriptor, which will be handled
-            # here as if directly passed
-            file_ = opener(file_, self._open_flags_from_open_modes(open_modes))
-
-        file_object, file_path, filedes, real_path = self._handle_file_arg(file_)
-        if file_object is None and file_path is None:
-            # file must be a fake pipe wrapper, find it...
-            if (
-                filedes is None
-                or len(self.filesystem.open_files) <= filedes
-                or not self.filesystem.open_files[filedes]
-            ):
-                raise OSError(errno.EBADF, "invalid pipe file descriptor")
-            wrappers = self.filesystem.open_files[filedes]
-            assert wrappers is not None
-            existing_wrapper = wrappers[0]
-            assert isinstance(existing_wrapper, FakePipeWrapper)
-            wrapper = FakePipeWrapper(
-                self.filesystem,
-                existing_wrapper.fd,
-                existing_wrapper.can_write,
-                mode,
-            )
-            file_des = self.filesystem._add_open_file(wrapper)
-            wrapper.filedes = file_des
-            return wrapper
-
-        assert file_path is not None
-        if not filedes:
-            closefd = True
-
-        if (
-            not opener
-            and open_modes.must_not_exist
-            and (
-                file_object
-                or self.filesystem.islink(file_path)
-                and not self.filesystem.is_windows_fs
-            )
-        ):
-            self.filesystem.raise_os_error(errno.EEXIST, file_path)
-
-        assert real_path is not None
-        file_object = self._init_file_object(
-            file_object, file_path, open_modes, real_path
-        )
-
-        if S_ISDIR(file_object.st_mode):
-            if self.filesystem.is_windows_fs:
-                self.filesystem.raise_os_error(errno.EACCES, file_path)
-            else:
-                self.filesystem.raise_os_error(errno.EISDIR, file_path)
-
-        # If you print obj.name, the argument to open() must be printed.
-        # Not the abspath, not the filename, but the actual argument.
-        file_object.opened_as = file_path
-        if open_modes.truncate:
-            current_time = now()
-            file_object.st_mtime = current_time
-            if not self.filesystem.is_windows_fs:
-                file_object.st_ctime = current_time
-
-        fakefile = FakeFileWrapper(
-            file_object,
-            file_path,
-            update=open_modes.can_write,
-            read=open_modes.can_read,
-            append=open_modes.append,
-            delete_on_close=self._delete_on_close,
-            filesystem=self.filesystem,
-            newline=newline,
-            binary=binary,
-            closefd=closefd,
-            encoding=encoding,
-            errors=errors,
-            buffering=buffering,
-            raw_io=self.raw_io,
-        )
-        if filedes is not None:
-            fakefile.filedes = filedes
-            # replace the file wrapper
-            open_files_list = self.filesystem.open_files[filedes]
-            assert open_files_list is not None
-            open_files_list.append(fakefile)
-        else:
-            fakefile.filedes = self.filesystem._add_open_file(fakefile)
-        return fakefile
-
-    @staticmethod
-    def _open_flags_from_open_modes(open_modes: _OpenModes) -> int:
-        flags = 0
-        if open_modes.can_read and open_modes.can_write:
-            flags |= os.O_RDWR
-        elif open_modes.can_read:
-            flags |= os.O_RDONLY
-        elif open_modes.can_write:
-            flags |= os.O_WRONLY
-
-        if open_modes.append:
-            flags |= os.O_APPEND
-        if open_modes.truncate:
-            flags |= os.O_TRUNC
-        if not open_modes.must_exist and open_modes.can_write:
-            flags |= os.O_CREAT
-        if open_modes.must_not_exist and open_modes.can_write:
-            flags |= os.O_EXCL
-        return flags
-
-    def _init_file_object(
-        self,
-        file_object: Optional[FakeFile],
-        file_path: AnyStr,
-        open_modes: _OpenModes,
-        real_path: AnyString,
-    ) -> FakeFile:
-        if file_object:
-            if not is_root() and (
-                (open_modes.can_read and not file_object.st_mode & PERM_READ)
-                or (open_modes.can_write and not file_object.st_mode & PERM_WRITE)
-            ):
-                self.filesystem.raise_os_error(errno.EACCES, file_path)
-            if open_modes.can_write:
-                if open_modes.truncate:
-                    file_object.set_contents("")
-        else:
-            if open_modes.must_exist:
-                self.filesystem.raise_os_error(errno.ENOENT, file_path)
-            if self.filesystem.islink(file_path):
-                link_object = self.filesystem.resolve(file_path, follow_symlinks=False)
-                assert link_object.contents is not None
-                target_path = cast(AnyStr, link_object.contents)
-            else:
-                target_path = file_path
-            if self.filesystem.ends_with_path_separator(target_path):
-                error = (
-                    errno.EINVAL
-                    if self.filesystem.is_windows_fs
-                    else errno.ENOENT
-                    if self.filesystem.is_macos
-                    else errno.EISDIR
-                )
-                self.filesystem.raise_os_error(error, file_path)
-            file_object = self.filesystem.create_file_internally(
-                real_path, create_missing_dirs=False, apply_umask=True
-            )
-        return file_object
-
-    def _handle_file_arg(
-        self, file_: Union[AnyStr, int]
-    ) -> Tuple[Optional[FakeFile], Optional[AnyStr], Optional[int], Optional[AnyStr]]:
-        file_object = None
-        if isinstance(file_, int):
-            # opening a file descriptor
-            filedes: int = file_
-            wrapper = self.filesystem.get_open_file(filedes)
-            if isinstance(wrapper, FakePipeWrapper):
-                return None, None, filedes, None
-            if isinstance(wrapper, FakeFileWrapper):
-                self._delete_on_close = wrapper.delete_on_close
-            file_object = cast(
-                FakeFile, self.filesystem.get_open_file(filedes).get_object()
-            )
-            assert file_object is not None
-            path = file_object.name
-            return file_object, cast(AnyStr, path), filedes, cast(AnyStr, path)
-
-        # open a file file by path
-        file_path = cast(AnyStr, file_)
-        if file_path == self.filesystem.dev_null.name:
-            file_object = self.filesystem.dev_null
-            real_path = file_path
-        else:
-            real_path = self.filesystem.resolve_path(file_path)
-            if self.filesystem.exists(file_path):
-                file_object = self.filesystem.get_object_from_normpath(
-                    real_path, check_read_perm=False
-                )
-        return file_object, file_path, None, real_path
-
-    def _handle_file_mode(
-        self,
-        mode: str,
-        newline: Optional[str],
-        open_modes: Optional[_OpenModes],
-    ) -> Tuple[Optional[str], _OpenModes]:
-        orig_modes = mode  # Save original modes for error messages.
-        # Normalize modes. Handle 't' and 'U'.
-        if ("b" in mode and "t" in mode) or (
-            sys.version_info > (3, 10) and "U" in mode
-        ):
-            raise ValueError("Invalid mode: " + mode)
-        mode = mode.replace("t", "").replace("b", "")
-        mode = mode.replace("rU", "r").replace("U", "r")
-        if not self.raw_io:
-            if mode not in _OPEN_MODE_MAP:
-                raise ValueError("Invalid mode: %r" % orig_modes)
-            open_modes = _OpenModes(*_OPEN_MODE_MAP[mode])
-        assert open_modes is not None
-        return newline, open_modes
 
 
 def _run_doctest() -> TestResults:
